Module graph::blockchain

The blockchain module exports the necessary traits and data structures to integrate a blockchain into Graph Node. A blockchain is represented by an implementation of the Blockchain trait which is the centerpiece of this module.

Re-exports

pub use block_stream::BlockStream;

pub use block_stream::ChainHeadUpdateListener;

pub use block_stream::ChainHeadUpdateStream;

pub use block_stream::TriggersAdapter;

Modules

block_ingestor

block_stream

types

Structs

BlockHash
A simple marker for byte arrays that are really block hashes

BlockPtr
A block hash and block number from a specific Ethereum block.

HostFn
Host fn that receives one u32 argument and returns an u32. The name for an AS fuction is in the format <namespace>.<function>.

HostFnCtx

Enums

IngestorError

Traits

Block

Blockchain

DataSource

DataSourceTemplate

IngestorAdapter

MappingTrigger

NodeCapabilities

RuntimeAdapter

TriggerData

TriggerFilter

UnresolvedDataSource

UnresolvedDataSourceTemplate

Type Definitions

BlockchainMap