Struct graph::blockchain::block_stream::BlockStreamContext

struct BlockStreamContext<C> where
    C: Blockchain,  {
    subgraph_store: Arc<dyn WritableStore>,
    chain_store: Arc<dyn ChainStore>,
    adapter: Arc<C::TriggersAdapter>,
    node_id: NodeId,
    subgraph_id: DeploymentHash,
    reorg_threshold: BlockNumber,
    filter: C::TriggerFilter,
    start_blocks: Vec<BlockNumber>,
    logger: Logger,
    metrics: Arc<BlockStreamMetrics>,
    previous_triggers_per_block: f64,
    previous_block_range_size: BlockNumber,
    max_block_range_size: BlockNumber,
    target_triggers_per_block_range: u64,
    unified_api_version: UnifiedMappingApiVersion,
}

Fields
subgraph_store: Arc<dyn WritableStore>chain_store: Arc<dyn ChainStore>adapter: Arc<C::TriggersAdapter>node_id: NodeIdsubgraph_id: DeploymentHashreorg_threshold: BlockNumberfilter: C::TriggerFilterstart_blocks: Vec<BlockNumber>logger: Loggermetrics: Arc<BlockStreamMetrics>previous_triggers_per_block: f64previous_block_range_size: BlockNumbermax_block_range_size: BlockNumbertarget_triggers_per_block_range: u64unified_api_version: UnifiedMappingApiVersion
Implementations

impl<C> BlockStreamContext<C> where
C: Blockchain,

async fn next_blocks(&self) -> Result<NextBlocks<C>, Error>

Perform reconciliation steps until there are blocks to yield or we are up-to-date.

async fn get_next_step(&self) -> Result<ReconciliationStep<C>, Error>

Determine the next reconciliation step. Does not modify Store or ChainStore.

fn update_subgraph_synced_status(&self) -> Result<(), Error>

Set subgraph deployment entity synced flag if and only if the subgraph block pointer is caught up to the head block pointer.