Trait graph::blockchain::block_stream::TriggersAdapter

pub trait TriggersAdapter<C: Blockchain>: Send + Sync {
    fn ancestor_block(
        &self, 
        ptr: BlockPtr, 
        offset: BlockNumber
    ) -> Result<Option<C::Block>, Error>;

#[must_use]    fn scan_triggers<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        from: BlockNumber, 
        to: BlockNumber, 
        filter: &'life1 C::TriggerFilter
    ) -> Pin<Box<dyn Future<Output = Result<Vec<BlockWithTriggers<C>>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn triggers_in_block<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        logger: &'life1 Logger, 
        block: C::Block, 
        filter: &'life2 C::TriggerFilter
    ) -> Pin<Box<dyn Future<Output = Result<BlockWithTriggers<C>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn is_on_main_chain<'life0, 'async_trait>(
        &'life0 self, 
        ptr: BlockPtr
    ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn parent_ptr<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        block: &'life1 BlockPtr
    ) -> Pin<Box<dyn Future<Output = Result<BlockPtr, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Required methods

fn ancestor_block(
&self,
ptr: BlockPtr,
offset: BlockNumber
) -> Result<Option<C::Block>, Error>

#[must_use]fn scan_triggers<'life0, 'life1, 'async_trait>(
&'life0 self,
from: BlockNumber,
to: BlockNumber,
filter: &'life1 C::TriggerFilter
) -> Pin<Box<dyn Future<Output = Result<Vec<BlockWithTriggers<C>>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

#[must_use]fn triggers_in_block<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
logger: &'life1 Logger,
block: C::Block,
filter: &'life2 C::TriggerFilter
) -> Pin<Box<dyn Future<Output = Result<BlockWithTriggers<C>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

#[must_use]fn is_on_main_chain<'life0, 'async_trait>(
&'life0 self,
ptr: BlockPtr
) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Return true if the block with the given hash and number is on the main chain, i.e., the chain going back from the current chain head.

#[must_use]fn parent_ptr<'life0, 'life1, 'async_trait>(
&'life0 self,
block: &'life1 BlockPtr
) -> Pin<Box<dyn Future<Output = Result<BlockPtr, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get pointer to parent of block. This is called when reverting block.

Loading content...