Struct graph::blockchain::BlockPtr

pub struct BlockPtr {
    pub hash: BlockHash,
    pub number: BlockNumber,
}

A block hash and block number from a specific Ethereum block.

Block numbers are signed 32 bit integers

Fields
hash: BlockHashnumber: BlockNumber
Implementations

impl BlockPtr

pub fn hash_hex(&self) -> String

Encodes the block hash into a hexadecimal string without a “0x” prefix. Hashes are stored in the database in this format.

pub fn block_number(&self) -> BlockNumber

Block number to be passed into the store. Panics if it does not fit in an i32.

pub fn hash_as_h256(&self) -> H256

pub fn hash_slice(&self) -> &[u8]