Trait graph::blockchain::Blockchain

pub trait Blockchain: Debug + Sized + Send + Sync + 'static {
    type Block: Block + Clone;
    type DataSource: DataSource<Self>;
    type UnresolvedDataSource: UnresolvedDataSource<Self>;
    type DataSourceTemplate: DataSourceTemplate<Self>;
    type UnresolvedDataSourceTemplate: UnresolvedDataSourceTemplate<Self>;
    type TriggersAdapter: TriggersAdapter<Self>;
    type TriggerData: TriggerData + Ord;
    type MappingTrigger: MappingTrigger + Debug;
    type TriggerFilter: TriggerFilter<Self>;
    type NodeCapabilities: NodeCapabilities<Self> + Display;
    type IngestorAdapter: IngestorAdapter<Self>;
    type RuntimeAdapter: RuntimeAdapter<Self>;
    fn reorg_threshold() -> u32;

    fn triggers_adapter(
        &self, 
        loc: &DeploymentLocator, 
        capabilities: &Self::NodeCapabilities, 
        unified_api_version: UnifiedMappingApiVersion, 
        stopwatch_metrics: StopwatchMetrics
    ) -> Result<Arc<Self::TriggersAdapter>, Error>;

    fn new_block_stream(
        &self, 
        deployment: DeploymentLocator, 
        start_blocks: Vec<BlockNumber>, 
        filter: Self::TriggerFilter, 
        metrics: Arc<BlockStreamMetrics>, 
        unified_api_version: UnifiedMappingApiVersion
    ) -> Result<BlockStream<Self>, Error>;

    fn ingestor_adapter(&self) -> Arc<Self::IngestorAdapter>;

    fn chain_store(&self) -> Arc<dyn ChainStore>;

#[must_use]    fn block_pointer_from_number<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        logger: &'life1 Logger, 
        number: BlockNumber
    ) -> Pin<Box<dyn Future<Output = Result<BlockPtr, IngestorError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn runtime_adapter(&self) -> Arc<Self::RuntimeAdapter>;
}

Associated Types

type Block: Block + Clone

type DataSource: DataSource<Self>

type UnresolvedDataSource: UnresolvedDataSource<Self>

type DataSourceTemplate: DataSourceTemplate<Self>

type UnresolvedDataSourceTemplate: UnresolvedDataSourceTemplate<Self>

type TriggersAdapter: TriggersAdapter<Self>

type TriggerData: TriggerData + Ord

Trigger data as parsed from the triggers adapter.

type MappingTrigger: MappingTrigger + Debug

Decoded trigger ready to be processed by the mapping.

type TriggerFilter: TriggerFilter<Self>

Trigger filter used as input to the triggers adapter.

type NodeCapabilities: NodeCapabilities<Self> + Display

type IngestorAdapter: IngestorAdapter<Self>

type RuntimeAdapter: RuntimeAdapter<Self>

Loading content...
Required methods

fn reorg_threshold() -> u32

fn triggers_adapter(
&self,
loc: &DeploymentLocator,
capabilities: &Self::NodeCapabilities,
unified_api_version: UnifiedMappingApiVersion,
stopwatch_metrics: StopwatchMetrics
) -> Result<Arc<Self::TriggersAdapter>, Error>

fn new_block_stream(
&self,
deployment: DeploymentLocator,
start_blocks: Vec<BlockNumber>,
filter: Self::TriggerFilter,
metrics: Arc<BlockStreamMetrics>,
unified_api_version: UnifiedMappingApiVersion
) -> Result<BlockStream<Self>, Error>

fn ingestor_adapter(&self) -> Arc<Self::IngestorAdapter>

fn chain_store(&self) -> Arc<dyn ChainStore>

#[must_use]fn block_pointer_from_number<'life0, 'life1, 'async_trait>(
&'life0 self,
logger: &'life1 Logger,
number: BlockNumber
) -> Pin<Box<dyn Future<Output = Result<BlockPtr, IngestorError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn runtime_adapter(&self) -> Arc<Self::RuntimeAdapter>

Loading content...