Trait graph::blockchain::IngestorAdapter

pub trait IngestorAdapter<C: Blockchain> {
    fn logger(&self) -> &Logger;

    fn ancestor_count(&self) -> BlockNumber;

#[must_use]    fn latest_block<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<BlockPtr, IngestorError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn ingest_block<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        hash: &'life1 BlockHash
    ) -> Pin<Box<dyn Future<Output = Result<Option<BlockHash>, IngestorError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn chain_head_ptr(&self) -> Result<Option<BlockPtr>, Error>;

    fn cleanup_cached_blocks(&self) -> Result<Option<(i32, usize)>, Error> { ... }
}

Required methods

fn logger(&self) -> &Logger

fn ancestor_count(&self) -> BlockNumber

How many ancestors of the current chain head to ingest. For chains that can experience reorgs, this should be large enough to cover all blocks that could be subject to reorgs to ensure that graph-node has enough blocks in its local cache to traverse a sidechain back to the main chain even if those blocks get removed from the network client.

#[must_use]fn latest_block<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<BlockPtr, IngestorError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the latest block from the chain

#[must_use]fn ingest_block<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 BlockHash
) -> Pin<Box<dyn Future<Output = Result<Option<BlockHash>, IngestorError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieve all necessary data for the block hash from the chain and store it in the database

fn chain_head_ptr(&self) -> Result<Option<BlockPtr>, Error>

Return the chain head that is stored locally, and therefore visible to the block streams of subgraphs

Loading content...
Provided methods

fn cleanup_cached_blocks(&self) -> Result<Option<(i32, usize)>, Error>

Remove old blocks from the database cache and return a pair containing the number of the oldest block retained and the number of blocks deleted if anything was removed. This is generally only used in small test installations, and can remain a noop without influencing correctness.

Loading content...