Trait graph::cheap_clone::CheapClone

pub trait CheapClone: Clone {
    fn cheap_clone(&self) -> Self { ... }
}

Things that are fast to clone in the context of an application such as Graph Node

The purpose of this API is to reduce the number of calls to .clone() which need to be audited for performance.

As a rule of thumb, only constant-time Clone impls should also implement CheapClone. Eg: ✔ Arc ✗ Vec ✔ u128 ✗ String

Provided methods

fn cheap_clone(&self) -> Self

Loading content...