Module graph::components::metrics

Components dealing with collecting metrics

Modules

aggregate
Aggregates over individual values.

stopwatch
Metrics for measuring where time is spent during indexing.

Macros

labels
Create labels with specified name-value pairs.

Structs

Histogram
A [Metric] counts individual observations from an event or sample stream in configurable buckets. Similar to a Summary, it also provides a sum of observations and an observation count.

HistogramOpts
A struct that bundles the options for creating a Histogram metric. It is mandatory to set Name and Help to a non-empty string. All other fields are optional and can safely be left at their zero value.

Opts
A struct that bundles the options for creating most [Metric] types.

Registry
A struct for registering Prometheus collectors, collecting their metrics, and gathering them into MetricFamilies for exposition.

Enums

PrometheusError
The error types for prometheus.

Traits

Collector
An interface for collecting metrics.

MetricsRegistry

Functions

counter_with_labels
Create an unregistered counter with labels

deployment_labels

gauge_with_labels
Create an unregistered gauge with labels

Type Definitions

Counter
A [Metric] represents a single numerical value that only ever goes up.

CounterVec
A Collector that bundles a set of Counters that all share the same [Desc], but have different values for their variable labels. This is used if you want to count the same thing partitioned by various dimensions (e.g. number of HTTP requests, partitioned by response code and method).

Gauge
A [Metric] represents a single numerical value that can arbitrarily go up and down.

GaugeVec
A Collector that bundles a set of Gauges that all share the same [Desc], but have different values for their variable labels. This is used if you want to count the same thing partitioned by various dimensions (e.g. number of operations queued, partitioned by user and operation type).

HistogramVec
A Collector that bundles a set of Histograms that all share the same [Desc], but have different values for their variable labels. This is used if you want to count the same thing partitioned by various dimensions (e.g. HTTP request latencies, partitioned by status code and method).