Struct graph::components::store::MockChainStore

pub struct MockChainStore {
    ChainStore_expectations: MockChainStore_ChainStore,
}

Common trait for blockchain store implementations.

Fields
ChainStore_expectations: MockChainStore_ChainStore
Implementations

impl MockChainStore

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockChainStore

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_genesis_block_ptr(&mut self) -> &mut Expectation

Create an Expectation for mocking the genesis_block_ptr method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_upsert_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the upsert_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_upsert_light_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the upsert_light_blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_attempt_chain_head_update(&mut self) -> &mut Expectation

Create an Expectation for mocking the attempt_chain_head_update method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_chain_head_ptr(&mut self) -> &mut Expectation

Create an Expectation for mocking the chain_head_ptr method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_ancestor_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the ancestor_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_cleanup_cached_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the cleanup_cached_blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_hashes_by_block_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_hashes_by_block_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_confirm_block_hash(&mut self) -> &mut Expectation

Create an Expectation for mocking the confirm_block_hash method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_transaction_receipts_in_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the transaction_receipts_in_block method