Struct graph::components::store::MockStore

pub struct MockStore {
    get_many_mock: Expectations,
}

Fields
get_many_mock: Expectations
Implementations

impl MockStore

pub fn get_many_mock<'a>(
&self,
_ids_for_type: BTreeMap<&'a EntityType, Vec<&'a str>>
) -> Result<BTreeMap<EntityType, Vec<Entity>>, StoreError>

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_get_many_mock(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_many_mock method

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new