Trait graph::components::store::ChainStore

pub trait ChainStore: Send + Sync + 'static {
    fn genesis_block_ptr(&self) -> Result<BlockPtr, Error>;

#[must_use]    fn upsert_block<'life0, 'async_trait>(
        &'life0 self, 
        block: EthereumBlock
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn upsert_light_blocks(
        &self, 
        blocks: Vec<LightEthereumBlock>
    ) -> Result<(), Error>;

#[must_use]    fn attempt_chain_head_update<'async_trait>(
        self: Arc<Self>, 
        ancestor_count: BlockNumber
    ) -> Pin<Box<dyn Future<Output = Result<Option<H256>, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait;

    fn chain_head_ptr(&self) -> Result<Option<BlockPtr>, Error>;

    fn blocks(
        &self, 
        hashes: Vec<H256>
    ) -> Result<Vec<LightEthereumBlock>, Error>;

    fn ancestor_block(
        &self, 
        block_ptr: BlockPtr, 
        offset: BlockNumber
    ) -> Result<Option<EthereumBlock>, Error>;

    fn cleanup_cached_blocks(
        &self, 
        ancestor_count: BlockNumber
    ) -> Result<Option<(BlockNumber, usize)>, Error>;

    fn block_hashes_by_block_number(
        &self, 
        number: BlockNumber
    ) -> Result<Vec<H256>, Error>;

    fn confirm_block_hash(
        &self, 
        number: BlockNumber, 
        hash: &H256
    ) -> Result<usize, Error>;

    fn block_number(
        &self, 
        block_hash: H256
    ) -> Result<Option<(String, BlockNumber)>, StoreError>;

#[must_use]    fn transaction_receipts_in_block<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        block_ptr: &'life1 H256
    ) -> Pin<Box<dyn Future<Output = Result<Vec<LightTransactionReceipt>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Common trait for blockchain store implementations.

Required methods

fn genesis_block_ptr(&self) -> Result<BlockPtr, Error>

Get a pointer to this blockchain’s genesis block.

#[must_use]fn upsert_block<'life0, 'async_trait>(
&'life0 self,
block: EthereumBlock
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Insert a block into the store (or update if they are already present).

fn upsert_light_blocks(
&self,
blocks: Vec<LightEthereumBlock>
) -> Result<(), Error>

#[must_use]fn attempt_chain_head_update<'async_trait>(
self: Arc<Self>,
ancestor_count: BlockNumber
) -> Pin<Box<dyn Future<Output = Result<Option<H256>, Error>> + Send + 'async_trait>> where
Self: 'async_trait,

Try to update the head block pointer to the block with the highest block number.

Only updates pointer if there is a block with a higher block number than the current head block, and the ancestor_count most recent ancestors of that block are in the store. Note that this means if the Ethereum node returns a different “latest block” with a different hash but same number, we do not update the chain head pointer. This situation can happen on e.g. Infura where requests are load balanced across many Ethereum nodes, in which case it’s better not to continuously revert and reapply the latest blocks.

If the pointer was updated, returns Ok(vec![]), and fires a HeadUpdateEvent.

If no block has a number higher than the current head block, returns Ok(vec![]).

If the candidate new head block had one or more missing ancestors, returns Ok(missing_blocks), where missing_blocks is a nonexhaustive list of missing blocks.

fn chain_head_ptr(&self) -> Result<Option<BlockPtr>, Error>

Get the current head block pointer for this chain. Any changes to the head block pointer will be to a block with a larger block number, never to a block with a smaller or equal block number.

The head block pointer will be None on initial set up.

fn blocks(&self, hashes: Vec<H256>) -> Result<Vec<LightEthereumBlock>, Error>

Returns the blocks present in the store.

fn ancestor_block(
&self,
block_ptr: BlockPtr,
offset: BlockNumber
) -> Result<Option<EthereumBlock>, Error>

Get the offsetth ancestor of block_hash, where offset=0 means the block matching block_hash and offset=1 means its parent. Returns None if unable to complete due to missing blocks in the chain store.

Returns an error if the offset would reach past the genesis block.

fn cleanup_cached_blocks(
&self,
ancestor_count: BlockNumber
) -> Result<Option<(BlockNumber, usize)>, Error>

Remove old blocks from the cache we maintain in the database and return a pair containing the number of the oldest block retained and the number of blocks deleted. We will never remove blocks that are within ancestor_count of the chain head.

fn block_hashes_by_block_number(
&self,
number: BlockNumber
) -> Result<Vec<H256>, Error>

Return the hashes of all blocks with the given number

fn confirm_block_hash(
&self,
number: BlockNumber,
hash: &H256
) -> Result<usize, Error>

Confirm that block number number has hash hash and that the store may purge any other blocks with that number

fn block_number(
&self,
block_hash: H256
) -> Result<Option<(String, BlockNumber)>, StoreError>

Find the block with block_hash and return the network name and number

#[must_use]fn transaction_receipts_in_block<'life0, 'life1, 'async_trait>(
&'life0 self,
block_ptr: &'life1 H256
) -> Pin<Box<dyn Future<Output = Result<Vec<LightTransactionReceipt>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Tries to retrieve all transactions receipts for a given block.

Loading content...