Trait graph::components::store::QueryStore

pub trait QueryStore: Send + Sync {
    fn find_query_values(
        &self, 
        query: EntityQuery
    ) -> Result<Vec<BTreeMap<String, Value>>, QueryExecutionError>;

#[must_use]    fn is_deployment_synced<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn block_ptr(&self) -> Result<Option<BlockPtr>, Error>;

    fn block_number(
        &self, 
        block_hash: H256
    ) -> Result<Option<BlockNumber>, StoreError>;

    fn wait_stats(&self) -> &PoolWaitStats;

#[must_use]    fn has_non_fatal_errors<'life0, 'async_trait>(
        &'life0 self, 
        block: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = Result<bool, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn deployment_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<DeploymentState, QueryExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn api_schema(&self) -> Result<Arc<ApiSchema>, QueryExecutionError>;

    fn network_name(&self) -> &str;

#[must_use]    fn query_permit<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = OwnedSemaphorePermit> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Store operations used when serving queries for a specific deployment

Required methods

fn find_query_values(
&self,
query: EntityQuery
) -> Result<Vec<BTreeMap<String, Value>>, QueryExecutionError>

#[must_use]fn is_deployment_synced<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn block_ptr(&self) -> Result<Option<BlockPtr>, Error>

fn block_number(
&self,
block_hash: H256
) -> Result<Option<BlockNumber>, StoreError>

fn wait_stats(&self) -> &PoolWaitStats

#[must_use]fn has_non_fatal_errors<'life0, 'async_trait>(
&'life0 self,
block: Option<BlockNumber>
) -> Pin<Box<dyn Future<Output = Result<bool, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

If block is None, assumes the latest block.

#[must_use]fn deployment_state<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<DeploymentState, QueryExecutionError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Find the current state for the subgraph deployment id and return details about it needed for executing queries

fn api_schema(&self) -> Result<Arc<ApiSchema>, QueryExecutionError>

fn network_name(&self) -> &str

#[must_use]fn query_permit<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = OwnedSemaphorePermit> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

A permit should be acquired before starting query execution.

Loading content...