Trait graph::components::store::QueryStoreManager

pub trait QueryStoreManager: Send + Sync + 'static {
#[must_use]    fn query_store<'life0, 'async_trait>(
        &'life0 self, 
        target: QueryTarget, 
        for_subscription: bool
    ) -> Pin<Box<dyn Future<Output = Result<Arc<dyn QueryStore + Send + Sync>, QueryExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required methods

#[must_use]fn query_store<'life0, 'async_trait>(
&'life0 self,
target: QueryTarget,
for_subscription: bool
) -> Pin<Box<dyn Future<Output = Result<Arc<dyn QueryStore + Send + Sync>, QueryExecutionError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get a new QueryStore. A QueryStore is tied to a DB replica, so if Graph Node is configured to use secondary DB servers the queries will be distributed between servers.

The query store is specific to a deployment, and id must indicate which deployment will be queried. It is not possible to use the id of the metadata subgraph, though the resulting store can be used to query metadata about the deployment id (but not metadata about other deployments).

If for_subscription is true, the main replica will always be used.

Loading content...