Trait graph::components::store::StatusStore

pub trait StatusStore: Send + Sync + 'static {
#[must_use]    fn query_permit<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = OwnedSemaphorePermit> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn status(&self, filter: Filter) -> Result<Vec<Info>, StoreError>;

    fn version_info(&self, version_id: &str) -> Result<VersionInfo, StoreError>;

    fn versions_for_subgraph_id(
        &self, 
        subgraph_id: &str
    ) -> Result<(Option<String>, Option<String>), StoreError>;

    fn get_proof_of_indexing<'a>(
        self: Arc<Self>, 
        subgraph_id: &'a DeploymentHash, 
        indexer: &'a Option<Address>, 
        block: BlockPtr
    ) -> DynTryFuture<'a, Option<[u8; 32]>>;
}

A view of the store that can provide information about the indexing status of any subgraph and any deployment

Required methods

#[must_use]fn query_permit<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = OwnedSemaphorePermit> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

A permit should be acquired before starting query execution.

fn status(&self, filter: Filter) -> Result<Vec<Info>, StoreError>

fn version_info(&self, version_id: &str) -> Result<VersionInfo, StoreError>

Support for the explorer-specific API

fn versions_for_subgraph_id(
&self,
subgraph_id: &str
) -> Result<(Option<String>, Option<String>), StoreError>

Support for the explorer-specific API; note that subgraph_id must be the id of an entry in subgraphs.subgraph, not that of a deployment. The return values are the ids of the subgraphs.subgraph_version for the current and pending versions of the subgraph

fn get_proof_of_indexing<'a>(
self: Arc<Self>,
subgraph_id: &'a DeploymentHash,
indexer: &'a Option<Address>,
block: BlockPtr
) -> DynTryFuture<'a, Option<[u8; 32]>>

A value of None indicates that the table is not available. Re-deploying the subgraph fixes this. It is undesirable to force everything to re-sync from scratch, so existing deployments will continue without a Proof of Indexing. Once all subgraphs have been re-deployed the Option can be removed.

Loading content...