Trait graph::components::store::SubgraphStore

pub trait SubgraphStore: Send + Sync + 'static {
    fn find_ens_name(
        &self, 
        _hash: &str
    ) -> Result<Option<String>, QueryExecutionError>;

    fn is_deployed(&self, id: &DeploymentHash) -> Result<bool, Error>;

    fn create_subgraph_deployment(
        &self, 
        name: SubgraphName, 
        schema: &Schema, 
        deployment: SubgraphDeploymentEntity, 
        node_id: NodeId, 
        network: String, 
        mode: SubgraphVersionSwitchingMode
    ) -> Result<DeploymentLocator, StoreError>;

    fn create_subgraph(&self, name: SubgraphName) -> Result<String, StoreError>;

    fn remove_subgraph(&self, name: SubgraphName) -> Result<(), StoreError>;

    fn reassign_subgraph(
        &self, 
        deployment: &DeploymentLocator, 
        node_id: &NodeId
    ) -> Result<(), StoreError>;

    fn assigned_node(
        &self, 
        deployment: &DeploymentLocator
    ) -> Result<Option<NodeId>, StoreError>;

    fn assignments(
        &self, 
        node: &NodeId
    ) -> Result<Vec<DeploymentLocator>, StoreError>;

    fn subgraph_exists(&self, name: &SubgraphName) -> Result<bool, StoreError>;

    fn input_schema(
        &self, 
        subgraph_id: &DeploymentHash
    ) -> Result<Arc<Schema>, StoreError>;

    fn api_schema(
        &self, 
        subgraph_id: &DeploymentHash
    ) -> Result<Arc<ApiSchema>, StoreError>;

    fn writable(
        &self, 
        deployment: &DeploymentLocator
    ) -> Result<Arc<dyn WritableStore>, StoreError>;

    fn writable_for_network_indexer(
        &self, 
        id: &DeploymentHash
    ) -> Result<Arc<dyn WritableStore>, StoreError>;

    fn least_block_ptr(
        &self, 
        id: &DeploymentHash
    ) -> Result<Option<BlockPtr>, Error>;

    fn locators(&self, hash: &str) -> Result<Vec<DeploymentLocator>, StoreError>;
}

Common trait for store implementations.

Required methods

fn find_ens_name(
&self,
_hash: &str
) -> Result<Option<String>, QueryExecutionError>

Find the reverse of keccak256 for hash through looking it up in the rainbow table.

fn is_deployed(&self, id: &DeploymentHash) -> Result<bool, Error>

Check if the store is accepting queries for the specified subgraph. May return true even if the specified subgraph is not currently assigned to an indexing node, as the store will still accept queries.

fn create_subgraph_deployment(
&self,
name: SubgraphName,
schema: &Schema,
deployment: SubgraphDeploymentEntity,
node_id: NodeId,
network: String,
mode: SubgraphVersionSwitchingMode
) -> Result<DeploymentLocator, StoreError>

Create a new deployment for the subgraph name. If the deployment already exists (as identified by the schema.id), reuse that, otherwise create a new deployment, and point the current or pending version of name at it, depending on the mode

fn create_subgraph(&self, name: SubgraphName) -> Result<String, StoreError>

Create a new subgraph with the given name. If one already exists, use the existing one. Return the id of the newly created or existing subgraph

fn remove_subgraph(&self, name: SubgraphName) -> Result<(), StoreError>

Remove a subgraph and all its versions; if deployments that were used by this subgraph do not need to be indexed anymore, also remove their assignment, but keep the deployments themselves around

fn reassign_subgraph(
&self,
deployment: &DeploymentLocator,
node_id: &NodeId
) -> Result<(), StoreError>

Assign the subgraph with id to the node node_id. If there is no assignment for the given deployment, report an error.

fn assigned_node(
&self,
deployment: &DeploymentLocator
) -> Result<Option<NodeId>, StoreError>

fn assignments(
&self,
node: &NodeId
) -> Result<Vec<DeploymentLocator>, StoreError>

fn subgraph_exists(&self, name: &SubgraphName) -> Result<bool, StoreError>

Return true if a subgraph name exists, regardless of whether the subgraph has any deployments attached to it

fn input_schema(
&self,
subgraph_id: &DeploymentHash
) -> Result<Arc<Schema>, StoreError>

Return the GraphQL schema supplied by the user

fn api_schema(
&self,
subgraph_id: &DeploymentHash
) -> Result<Arc<ApiSchema>, StoreError>

Return the GraphQL schema that was derived from the user’s schema by adding a root query type etc. to it

fn writable(
&self,
deployment: &DeploymentLocator
) -> Result<Arc<dyn WritableStore>, StoreError>

Return a WritableStore that is used for indexing subgraphs. Only code that is part of indexing a subgraph should ever use this.

fn writable_for_network_indexer(
&self,
id: &DeploymentHash
) -> Result<Arc<dyn WritableStore>, StoreError>

The network indexer does not follow the normal flow of how subgraphs are indexed, and therefore needs a special way to get a WritableStore. This method should not be used outside of that, and writable should be used instead

fn least_block_ptr(
&self,
id: &DeploymentHash
) -> Result<Option<BlockPtr>, Error>

Return the minimum block pointer of all deployments with this id that we would use to query or copy from; in particular, this will ignore any instances of this deployment that are in the process of being set up

fn locators(&self, hash: &str) -> Result<Vec<DeploymentLocator>, StoreError>

Find the deployment locators for the subgraph with the given hash

Loading content...