Trait graph::components::store::WritableStore

pub trait WritableStore: Send + Sync + 'static {
    fn block_ptr(&self) -> Result<Option<BlockPtr>, Error>;

    fn start_subgraph_deployment(
        &self, 
        logger: &Logger
    ) -> Result<(), StoreError>;

    fn revert_block_operations(
        &self, 
        block_ptr_to: BlockPtr
    ) -> Result<(), StoreError>;

    fn unfail(&self) -> Result<(), StoreError>;

#[must_use]    fn fail_subgraph<'life0, 'async_trait>(
        &'life0 self, 
        error: SubgraphError
    ) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn supports_proof_of_indexing<'a>(self: Arc<Self>) -> DynTryFuture<'a, bool>;

    fn get(&self, key: EntityKey) -> Result<Option<Entity>, QueryExecutionError>;

    fn transact_block_operations(
        &self, 
        block_ptr_to: BlockPtr, 
        mods: Vec<EntityModification>, 
        stopwatch: StopwatchMetrics, 
        data_sources: Vec<StoredDynamicDataSource>, 
        deterministic_errors: Vec<SubgraphError>
    ) -> Result<(), StoreError>;

    fn get_many(
        &self, 
        ids_for_type: BTreeMap<&EntityType, Vec<&str>>
    ) -> Result<BTreeMap<EntityType, Vec<Entity>>, StoreError>;

    fn deployment_synced(&self) -> Result<(), Error>;

#[must_use]    fn is_deployment_synced<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn unassign_subgraph(&self) -> Result<(), StoreError>;

#[must_use]    fn load_dynamic_data_sources<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredDynamicDataSource>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required methods

fn block_ptr(&self) -> Result<Option<BlockPtr>, Error>

Get a pointer to the most recently processed block in the subgraph.

fn start_subgraph_deployment(&self, logger: &Logger) -> Result<(), StoreError>

Start an existing subgraph deployment.

fn revert_block_operations(
&self,
block_ptr_to: BlockPtr
) -> Result<(), StoreError>

Revert the entity changes from a single block atomically in the store, and update the subgraph block pointer to block_ptr_to.

block_ptr_to must point to the parent block of the subgraph block pointer.

fn unfail(&self) -> Result<(), StoreError>

Remove the fatal error from a subgraph and check if it is healthy or unhealthy.

#[must_use]fn fail_subgraph<'life0, 'async_trait>(
&'life0 self,
error: SubgraphError
) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Set subgraph status to failed with the given error as the cause.

fn supports_proof_of_indexing<'a>(self: Arc<Self>) -> DynTryFuture<'a, bool>

fn get(&self, key: EntityKey) -> Result<Option<Entity>, QueryExecutionError>

Looks up an entity using the given store key at the latest block.

fn transact_block_operations(
&self,
block_ptr_to: BlockPtr,
mods: Vec<EntityModification>,
stopwatch: StopwatchMetrics,
data_sources: Vec<StoredDynamicDataSource>,
deterministic_errors: Vec<SubgraphError>
) -> Result<(), StoreError>

Transact the entity changes from a single block atomically into the store, and update the subgraph block pointer to block_ptr_to.

block_ptr_to must point to a child block of the current subgraph block pointer.

fn get_many(
&self,
ids_for_type: BTreeMap<&EntityType, Vec<&str>>
) -> Result<BTreeMap<EntityType, Vec<Entity>>, StoreError>

Look up multiple entities as of the latest block. Returns a map of entities by type.

fn deployment_synced(&self) -> Result<(), Error>

The deployment id finished syncing, mark it as synced in the database and promote it to the current version in the subgraphs where it was the pending version so far

#[must_use]fn is_deployment_synced<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Return true if the deployment with the given id is fully synced, and return false otherwise. Errors from the store are passed back up

fn unassign_subgraph(&self) -> Result<(), StoreError>

#[must_use]fn load_dynamic_data_sources<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredDynamicDataSource>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load the dynamic data sources for the given deployment

Loading content...