Module graph::components::subgraph

Components dealing with subgraphs.

Re-exports

pub use crate::prelude::Entity;

Modules

host

instance

instance_manager

proof_of_indexing

provider

registrar

Structs

BlockEventStream

BlockState

DataSourceTemplateInfo

HostMetrics

ProofOfIndexing

ProofOfIndexingFinisher

Enums

MappingError

ProofOfIndexingEvent

SubgraphVersionSwitchingMode

Traits

RuntimeHost
Common trait for runtime host implementations.

RuntimeHostBuilder

SubgraphAssignmentProvider
Common trait for subgraph providers.

SubgraphInstanceManager
A SubgraphInstanceManager loads and manages subgraph instances.

SubgraphRegistrar
Common trait for subgraph registrars.

Type Definitions

SharedProofOfIndexing
This concoction of types is to allow MappingContext to be static, yet still have shared mutable data for derive_with_empty_block_state. The static requirement is so that host exports can be static for wasmtime. AtomicRefCell is chosen over Mutex because concurrent access is intentionally disallowed - PoI requires sequential access to the hash function within a given causality region even if ownership is shared across multiple mapping contexts.