Module graph::components::subgraph::proof_of_indexing

Re-exports

pub use event::ProofOfIndexingEvent;

pub use online::BlockEventStream;

pub use online::ProofOfIndexing;

pub use online::ProofOfIndexingFinisher;

Modules

event

online
This is an online (streaming) implementation of the reference implementation Any hash constructed from here should be the same as if the same data was given to the reference implementation, but this is updated incrementally

reference

Type Definitions

SharedProofOfIndexing
This concoction of types is to allow MappingContext to be static, yet still have shared mutable data for derive_with_empty_block_state. The static requirement is so that host exports can be static for wasmtime. AtomicRefCell is chosen over Mutex because concurrent access is intentionally disallowed - PoI requires sequential access to the hash function within a given causality region even if ownership is shared across multiple mapping contexts.