Type Definition graph::components::subgraph::proof_of_indexing::SharedProofOfIndexing

type SharedProofOfIndexing = Option<Arc<AtomicRefCell<ProofOfIndexing>>>;

This concoction of types is to allow MappingContext to be static, yet still have shared mutable data for derive_with_empty_block_state. The static requirement is so that host exports can be static for wasmtime. AtomicRefCell is chosen over Mutex because concurrent access is intentionally disallowed - PoI requires sequential access to the hash function within a given causality region even if ownership is shared across multiple mapping contexts.

The Option<_> is because not all subgraphs support PoI until re-deployed. Eventually this can be removed.

This is not a great place to define this type, since the ProofOfIndexing shouldn’t “know” these details about wasmtime and subgraph re-deployments, but the APIs that would make use of this are in graph/components so this lives here for lack of a better choice.