Trait graph::components::subgraph::RuntimeHostBuilder

pub trait RuntimeHostBuilder<C: Blockchain>: Clone + Send + Sync + 'static {
    type Host: RuntimeHost<C> + PartialEq;
    type Req: 'static + Send;
    fn build(
        &self, 
        network_name: String, 
        subgraph_id: DeploymentHash, 
        data_source: C::DataSource, 
        top_level_templates: Arc<Vec<C::DataSourceTemplate>>, 
        mapping_request_sender: Sender<Self::Req>, 
        metrics: Arc<HostMetrics>
    ) -> Result<Self::Host, Error>;

    fn spawn_mapping(
        raw_module: Vec<u8>, 
        logger: Logger, 
        subgraph_id: DeploymentHash, 
        metrics: Arc<HostMetrics>
    ) -> Result<Sender<Self::Req>, Error>;
}

Associated Types

type Host: RuntimeHost<C> + PartialEq

type Req: 'static + Send

Loading content...
Required methods

fn build(
&self,
network_name: String,
subgraph_id: DeploymentHash,
data_source: C::DataSource,
top_level_templates: Arc<Vec<C::DataSourceTemplate>>,
mapping_request_sender: Sender<Self::Req>,
metrics: Arc<HostMetrics>
) -> Result<Self::Host, Error>

Build a new runtime host for a subgraph data source.

fn spawn_mapping(
raw_module: Vec<u8>,
logger: Logger,
subgraph_id: DeploymentHash,
metrics: Arc<HostMetrics>
) -> Result<Sender<Self::Req>, Error>

Spawn a mapping and return a channel for mapping requests. The sender should be able to be cached and shared among mappings that use the same wasm file.

Loading content...