Module graph::components::transaction_receipt

Code for retrieving transaction receipts from the database.

This module exposes:

the find_transaction_receipts_in_block function, that queries the database and returns transaction receipts present in a given block.

the LightTransactionReceipt type, which holds basic information about the retrieved transaction receipts.

Structs

LightTransactionReceipt
Like web3::types::Receipt, but with fewer fields.

RawTransactionReceipt
Type that comes straight out of a SQL query

TransactionReceiptQuery
Parameters for querying for all transaction receipts of a given block.

Functions

_impl_queryable_by_name_for_rawtransactionreceipt

_impl_queryable_for_rawtransactionreceipt

drain_vector
Converts Vec to [u8; N], where N is the vector’s expected lenght. Fails if input size is larger than output size.

find_transaction_receipts_in_block
Queries the database for all the transaction receipts in a given block range.