Module graph::data::graphql

Data types for dealing with GraphQL values.

Re-exports

pub use ext::DirectiveExt;

pub use ext::DocumentExt;

pub use ext::ObjectTypeExt;

pub use ext::TypeExt;

pub use ext::ValueExt;

pub use object_or_interface::ObjectOrInterface;

pub use crate::object;

pub use object_macro::object_value;

pub use object_macro::IntoValue;

Modules

effort
Utilities to keep moving statistics about queries

ext
Traits to navigate the GraphQL AST

object_macro

object_or_interface

serialization

shape_hash
Calculate a hash for a GraphQL query that reflects the shape of the query. The shape hash will be the same for two instances of a query that are deemed identical except for unimportant details. Those details are any values used with filters, and any differences in the query name or response keys

values
Utilities for working with GraphQL values.

Structs

SerializableValue
Serializable wrapper around a GraphQL value.

Traits

TryFromValue

ValueList

ValueMap