Module graph::data::graphql::effort

Utilities to keep moving statistics about queries

Structs

JAIL_QUERIES

JAIL_THRESHOLD

KillState

LOAD_MANAGEMENT_DISABLED

LOAD_THRESHOLD

LoadManager

QueryEffort

QueryEffortInner
Track the effort for queries (identified by their ShapeHash) over a time window.

SIMULATE

Enums

Decision
Indicate what the load manager wants query execution to do with a query

KillStateLogEvent
What to log about the state we are currently in

Constants

ZERO_DURATION