Struct graph::data::schema::Schema

pub struct Schema {
    pub id: DeploymentHash,
    pub document: Document,
    pub interfaces_for_type: BTreeMap<EntityType, Vec<InterfaceType>>,
    pub types_for_interface: BTreeMap<EntityType, Vec<ObjectType>>,
}

A validated and preprocessed GraphQL schema for a subgraph.

Fields
id: DeploymentHashdocument: Documentinterfaces_for_type: BTreeMap<EntityType, Vec<InterfaceType>>types_for_interface: BTreeMap<EntityType, Vec<ObjectType>>
Implementations

impl Schema

pub fn new(id: DeploymentHash, document: Document) -> Self

Create a new schema. The document must already have been validated. This function is only useful for creating an introspection schema, and should not be used otherwise

pub fn resolve_schema_references<S: SubgraphStore>(
&self,
store: Arc<S>
) -> (HashMap<SchemaReference, Arc<Schema>>, Vec<SchemaImportError>)

fn resolve_import_graph<S: SubgraphStore>(
&self,
store: Arc<S>,
schemas: &mut HashMap<SchemaReference, Arc<Schema>>,
visit_log: &mut HashSet<DeploymentHash>
) -> Vec<SchemaImportError>

pub fn collect_interfaces(
document: &Document
) -> Result<(BTreeMap<EntityType, Vec<InterfaceType>>, BTreeMap<EntityType, Vec<ObjectType>>), SchemaValidationError>

pub fn parse(raw: &str, id: DeploymentHash) -> Result<Self, Error>

fn imported_types(&self) -> HashMap<ImportedType, SchemaReference>

pub fn imported_schemas(&self) -> Vec<SchemaReference>

pub fn name_argument_value_from_directive(directive: &Directive) -> Value

pub fn types_for_interface(&self) -> &BTreeMap<EntityType, Vec<ObjectType>>

Returned map has one an entry for each interface in the schema.

pub fn interfaces_for_type(
&self,
type_name: &EntityType
) -> Option<&Vec<InterfaceType>>

Returns None if the type implements no interfaces.

pub fn add_subgraph_id_directives(&mut self, id: DeploymentHash)

pub fn validate(
&self,
schemas: &HashMap<SchemaReference, Arc<Schema>>
) -> Result<(), Vec<SchemaValidationError>>

fn validate_schema_type_has_no_fields(
&self
) -> Result<(), SchemaValidationError>

fn validate_directives_on_schema_type(
&self
) -> Result<(), SchemaValidationError>

fn validate_import_directive_arguments(
import: &Directive
) -> Option<SchemaValidationError>

Check the syntax of a single @import directive

fn validate_import_directive_schema_reference_parses(
directive: &Directive
) -> Option<SchemaValidationError>

fn validate_fulltext_directives(&self) -> Vec<SchemaValidationError>

fn validate_fulltext_directive_name(
&self,
fulltext: &Directive
) -> Vec<SchemaValidationError>

fn validate_fulltext_directive_language(
&self,
fulltext: &Directive
) -> Vec<SchemaValidationError>

fn validate_fulltext_directive_algorithm(
&self,
fulltext: &Directive
) -> Vec<SchemaValidationError>

fn validate_fulltext_directive_includes(
&self,
fulltext: &Directive
) -> Vec<SchemaValidationError>

fn validate_import_directives(&self) -> Vec<SchemaValidationError>

fn validate_imported_types(
&self,
schemas: &HashMap<SchemaReference, Arc<Schema>>
) -> Vec<SchemaValidationError>

fn validate_fields(&self) -> Vec<SchemaValidationError>

fn validate_reserved_types_usage(&self) -> Result<(), SchemaValidationError>

Checks if the schema is using types that are reserved by graph-node

fn validate_schema_types(&self) -> Result<(), SchemaValidationError>

fn validate_derived_from(&self) -> Result<(), SchemaValidationError>

fn validate_interface_implementation(
object: &ObjectType,
interface: &InterfaceType
) -> Result<(), SchemaValidationError>

Validate that object implements interface.

fn subgraph_schema_object_type(&self) -> Option<&ObjectType>

pub fn entity_fulltext_definitions<'a>(
entity: &str,
document: &'a Document
) -> Result<Vec<FulltextDefinition>, Error>