Module graph::data::store

Data types for dealing with storing entities.

Modules

ethereum

scalar
Custom scalars in GraphQL.

Structs

Entity
An entity is represented as a map of attribute names to values.

NodeId

Enums

AssignmentEvent

SubscriptionFilter
Filter subscriptions

Value
An attribute value is represented as an enum with variants for all supported value types.

ValueType

Constants

BIG_DECIMAL_SCALAR

BIG_INT_SCALAR

BYTES_SCALAR

ID

Traits

ToEntityId
A value that can be converted to an Entity ID.

ToEntityKey
A value that can be converted to an Entity key.

TryIntoEntity
A value that can (maybe) be converted to an Entity.

Type Definitions

Attribute
An entity attribute name is represented as a string.