Struct graph::data::store::Entity

pub struct Entity(HashMap<Attribute, Value>);

An entity is represented as a map of attribute names to values.

Implementations

impl Entity

pub fn new() -> Self

Creates a new entity with no attributes set.

pub fn get(&self, key: &str) -> Option<&Value>

pub fn insert(&mut self, key: String, value: Value) -> Option<Value>

pub fn remove(&mut self, key: &str) -> Option<Value>

pub fn contains_key(&mut self, key: &str) -> bool

pub fn sorted(self) -> Vec<(String, Value)>

pub fn id(&self) -> Result<String, Error>

Try to get this entity’s ID

pub fn set(
&mut self,
name: impl Into<Attribute>,
value: impl Into<Value>
) -> Option<Value>

Convenience method to save having to .into() the arguments.

pub fn merge(&mut self, update: Entity)

Merges an entity update update into this entity.

If a key exists in both entities, the value from update is chosen. If a key only exists on one entity, the value from that entity is chosen. If a key is set to Value::Null in update, the key/value pair is set to Value::Null.

pub fn merge_remove_null_fields(&mut self, update: Entity)

Merges an entity update update into this entity, removing Value::Null values.

If a key exists in both entities, the value from update is chosen. If a key only exists on one entity, the value from that entity is chosen. If a key is set to Value::Null in update, the key/value pair is removed.