Module graph::data::subgraph

Data types for dealing with subgraphs.

Modules

schema
Rust representation of the GraphQL schema for a SubgraphManifest. Entity types that contain the graph-node state.

status
Support for the indexing status API

Structs

BaseSubgraphManifest

CreateSubgraphResult
Result of a creating a subgraph in the registar.

DISABLE_GRAFTS

DeploymentHash
The IPFS hash used to identifiy a deployment externally, i.e., the Qm.. string that graph-cli prints when deploying to a subgraph

DeploymentState
Important details about the current state of a subgraph deployment used while executing queries against a deployment

DifferentMappingApiVersions

Graft

Link
IPLD link.

MAX_API_VERSION

MAX_SPEC_VERSION

MIN_SPEC_VERSION

Mapping

MappingABI

MappingBlockHandler

MappingCallHandler

MappingEventHandler

Source

SubgraphName

TemplateSource

UnifiedMappingApiVersion

UnresolvedMapping

UnresolvedMappingABI

UnresolvedSchema

UnvalidatedSubgraphManifest
Unvalidated SubgraphManifest

Enums

BlockHandlerFilter

SubgraphAssignmentProviderError

SubgraphFeature

SubgraphManifestResolveError

SubgraphManifestValidationError

SubgraphManifestValidationWarning

SubgraphRegistrarError

Constants

API_VERSION_0_0_5
This version adds a new validation step that rejects manifests whose mappings have different API versions if at least one of them is equal to or higher than 0.0.5.

Functions

deserialize_address
Deserialize an Address (with or without ‘0x’ prefix).

display_vector

format_versions

Type Definitions

DataSourceContext
Data source contexts are conveniently represented as entities.

SubgraphManifest
SubgraphManifest validated with IPFS links resolved

UnresolvedSubgraphManifest
SubgraphManifest with IPFS links unresolved