Module graph::dijkstra

Main class for performing Dijkstra's algorithm.

The idea of this class is to process a graph and yield all its elements in order of distance.

Examples

use graph::Graph;
//The actual way to create such an iterator is through the 'GraphTrait' trait;
use graph::GraphTrait;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

let mut iter = graph.gen_limited_dijkstra_vec(&vec![0], |edge| edge.len(), |_, _| true);

// Node 0 is at distance 0 (obviously)
assert_eq!(iter.next(), Some((0, 0)));
// Node 5 is at distance "Edge from A to B".len() = 16
assert_eq!(iter.next(), Some((5, 16)));

Additionally the route to the source can be queried using the "Visited" field:

// Lock the structure on the stack
let visited = iter.visited();
// Borrow the structure
let visited = visited.borrow();

// The node 5 is visited via 0, so the shortest path from 0 to 5 is through 0.
assert_eq!(visited.get(&5), Some(&Some(0)));
// 0 is the origin, so it has no previous node.
assert_eq!(visited.get(&0), Some(&None));
// 3 is not visited.
assert_eq!(visited.get(&3), None);

Structs

DijkstraGenerator
Iterator for generating all edges and nodes in a graph, in a Dijkstra's fashion.

Type Definitions

Visited
Shared linked hashmap for a path.