Trait graph::ext::futures::FutureExtension

pub trait FutureExtension: Future + Sized {
    fn cancelable<C: Fn() -> Self::Error>(
        self, 
        guard: &impl Canceler, 
        on_cancel: C
    ) -> Cancelable<Self, C>;

    fn timeout(self, dur: Duration) -> Timeout<Compat01As03<Self>>;
}

Required methods

fn cancelable<C: Fn() -> Self::Error>(
self,
guard: &impl Canceler,
on_cancel: C
) -> Cancelable<Self, C>

When cancel is called on a CancelGuard or it is dropped, Cancelable receives an error.

on_cancel is called to make an error value upon cancelation.

fn timeout(self, dur: Duration) -> Timeout<Compat01As03<Self>>

Loading content...