Trait graph::ext::futures::StreamExtension

pub trait StreamExtension: Stream + Sized {
    fn cancelable<C: Fn() -> Self::Error>(
        self, 
        guard: &impl Canceler, 
        on_cancel: C
    ) -> Cancelable<Self, C>;
}

Required methods

fn cancelable<C: Fn() -> Self::Error>(
self,
guard: &impl Canceler,
on_cancel: C
) -> Cancelable<Self, C>

When cancel is called on a CancelGuard or it is dropped, Cancelable receives an error.

on_cancel is called to make an error value upon cancelation.

Loading content...