Module graph::graphs

This module contains default implementations for graphs and edges

Structs

UndirectedAdjacencyListGraph
An undirected graph represented by an adjacency list.

UnweightedEdge
A default implementation of a weighted edge that can be used in graph implementations.

WeightedEdge
A default implementation of a weighted edge that can be used in graph implementations.