Struct graph::matrix_graph::MatrixGraph

pub struct MatrixGraph<N, T> where
    N: MatrixGraphNode, { /* private fields */ }

Graph representation with adjacency matrix

Note: it’s better to use for dense graph

Implementations

impl<N, T> MatrixGraph<N, T> where
N: MatrixGraphNode,

pub fn from_edges(edges: impl IntoIterator<Item = (N, N, T)>) -> Self

Create MatrixGraph from iterator of tuples where each element representes edge between two nodes and it’s weight

impl<N, T> MatrixGraph<N, T> where
N: MatrixGraphNode,

pub fn add_node(&mut self, node: N) -> usize

Adds new node in graph

Returns index of new node

Computes in O(1) (average amortized) Worse case O(n) where n is nodes count

Panics if node already exists

pub fn remove_node(&mut self, node_index: usize) -> Option<N>

Removes node and all edges for it

Returns removed node or None, if node not found

Computes in O(e) (average) where e = node’s edges count

pub fn add_edge(&mut self, from: usize, to: usize, weight: T)

Adds edge between two nodes

Computes in O(1) (average) Worst case O(n ^ 2) where n = nodes count

Panics if some of nodes not exists or edge already exists

pub fn remove_edge(&mut self, from_node: usize, to_node: usize) -> Option<T>

Removes edge between two nodes

Returns edge’s weight if removed else None

Computes in O(1)

pub fn node_count(&self) -> usize

Returns count of nodes

Computes in O(1)

pub fn edge_count(&self) -> usize

Returns count of edges

Computes in O(1)

pub fn node_iter(&mut self) -> NodeIterator<'_, N>

Returns iterator over nodes

pub fn contains_node(&self, node: &N) -> bool

Checks if node exists in graph

Computes in O(1) (average)

pub fn contains_edge(&self, from: usize, to: usize) -> bool

Checks if edge between two nodes exists

Computes in O(1)

pub fn get_index_of(&self, node: &N) -> Option<usize>

Returns index of node or None if not found

Computes in O(1) (average)