Module graph::nauty

Module containg methods to call the nauty library in order to compute the canonical form or the orbits of the vertices of a graph.

Functions

canon_graph
Given a graph, returns the canonical form of the graph, the order of the vertices of g in the new graph and the orbits with same format as nauty. The return values are in this order.

canon_graph_fixed
Given a graph and a set of fixed vertices, returns the canonical form of the graph, the order of the vertices of g in the new graph and the orbits with same format as nauty. The return values are in this order. Example:

canon_graph_fixed_directed

canon_transfo

canon_transfo_with_fixed

orbits
Returns a list containing one vertex per orbit. Example: