Module graph::prelude

A prelude that makes all system component traits and data types available.

Add the following code to import all traits and data types listed below at once.

use graph::prelude::*;

Re-exports

pub use super::entity;

pub use anyhow;

pub use anyhow::Context as _;

pub use bigdecimal;

pub use chrono;

pub use ethabi;

pub use futures03;

pub use futures03::future::FutureExt as _;

pub use futures03::sink::SinkExt as _;

pub use futures03::stream::StreamExt as _;

pub use hex;

pub use rand;

pub use reqwest;

pub use serde;

pub use serde_json;

pub use serde_yaml;

pub use slog;

pub use thiserror;

pub use tiny_keccak;

pub use tokio;

pub use web3;

pub use crate::components::ethereum::EthereumNetworkIdentifier;

pub use crate::components::graphql::GraphQlRunner;

pub use crate::components::graphql::QueryLoadManager;

pub use crate::components::graphql::SubscriptionResultFuture;

pub use crate::components::link_resolver::JsonStreamValue;

pub use crate::components::link_resolver::JsonValueStream;

pub use crate::components::link_resolver::LinkResolver;

pub use crate::components::metrics::aggregate::Aggregate;

pub use crate::components::metrics::stopwatch::StopwatchMetrics;

pub use crate::components::metrics::MetricsRegistry;

pub use crate::components::server::admin::JsonRpcServer;

pub use crate::components::server::index_node::IndexNodeServer;

pub use crate::components::server::metrics::MetricsServer;

pub use crate::components::server::query::GraphQLServer;

pub use crate::components::server::subscription::SubscriptionServer;

pub use crate::components::store::AttributeNames;

pub use crate::components::store::BlockNumber;

pub use crate::components::store::ChainStore;

pub use crate::components::store::ChildMultiplicity;

pub use crate::components::store::EntityCache;

pub use crate::components::store::EntityChange;

pub use crate::components::store::EntityChangeOperation;

pub use crate::components::store::EntityCollection;

pub use crate::components::store::EntityFilter;

pub use crate::components::store::EntityKey;

pub use crate::components::store::EntityLink;

pub use crate::components::store::EntityModification;

pub use crate::components::store::EntityOperation;

pub use crate::components::store::EntityOrder;

pub use crate::components::store::EntityQuery;

pub use crate::components::store::EntityRange;

pub use crate::components::store::EntityWindow;

pub use crate::components::store::EthereumCallCache;

pub use crate::components::store::ParentLink;

pub use crate::components::store::PoolWaitStats;

pub use crate::components::store::QueryStore;

pub use crate::components::store::QueryStoreManager;

pub use crate::components::store::StoreError;

pub use crate::components::store::StoreEvent;

pub use crate::components::store::StoreEventStream;

pub use crate::components::store::StoreEventStreamBox;

pub use crate::components::store::SubgraphStore;

pub use crate::components::store::WindowAttribute;

pub use crate::components::store::BLOCK_NUMBER_MAX;

pub use crate::components::store::SUBSCRIPTION_THROTTLE_INTERVAL;

pub use crate::components::store::SUBSCRIPTION_THROTTLE_INTERVAL;

pub use crate::components::transaction_receipt;

pub use crate::components::EventConsumer;

pub use crate::components::EventProducer;

pub use crate::cheap_clone::CheapClone;

pub use crate::data::graphql::shape_hash::shape_hash;

pub use crate::data::schema::ApiSchema;

pub use crate::data::schema::Schema;

pub use crate::data::store::ethereum::*;

pub use crate::data::store::scalar::BigDecimal;

pub use crate::data::store::scalar::BigInt;

pub use crate::data::store::AssignmentEvent;

pub use crate::data::store::Attribute;

pub use crate::data::store::Entity;

pub use crate::data::store::NodeId;

pub use crate::data::store::SubscriptionFilter;

pub use crate::data::store::ToEntityId;

pub use crate::data::store::ToEntityKey;

pub use crate::data::store::TryIntoEntity;

pub use crate::data::store::Value;

pub use crate::data::store::ValueType;

pub use crate::data::subgraph::schema::SubgraphDeploymentEntity;

pub use crate::data::subgraph::BlockHandlerFilter;

pub use crate::data::subgraph::CreateSubgraphResult;

pub use crate::data::subgraph::DataSourceContext;

pub use crate::data::subgraph::DeploymentHash;

pub use crate::data::subgraph::DeploymentState;

pub use crate::data::subgraph::Link;

pub use crate::data::subgraph::MappingABI;

pub use crate::data::subgraph::MappingBlockHandler;

pub use crate::data::subgraph::MappingCallHandler;

pub use crate::data::subgraph::MappingEventHandler;

pub use crate::data::subgraph::SubgraphAssignmentProviderError;

pub use crate::data::subgraph::SubgraphManifest;

pub use crate::data::subgraph::SubgraphManifestResolveError;

pub use crate::data::subgraph::SubgraphManifestValidationError;

pub use crate::data::subgraph::SubgraphName;

pub use crate::data::subgraph::SubgraphRegistrarError;

pub use crate::data::subgraph::UnvalidatedSubgraphManifest;

pub use crate::ext::futures::CancelGuard;

pub use crate::ext::futures::CancelHandle;

pub use crate::ext::futures::CancelToken;

pub use crate::ext::futures::CancelableError;

pub use crate::ext::futures::FutureExtension;

pub use crate::ext::futures::SharedCancelGuard;

pub use crate::ext::futures::StreamExtension;

pub use crate::impl_slog_value;

pub use crate::log::codes::LogCode;

pub use crate::log::elastic::elastic_logger;

pub use crate::log::elastic::ElasticDrainConfig;

pub use crate::log::elastic::ElasticLoggingConfig;

pub use crate::log::factory::ComponentLoggerConfig;

pub use crate::log::factory::ElasticComponentLoggerConfig;

pub use crate::log::factory::LoggerFactory;

pub use crate::log::split::split_logger;

pub use crate::util::cache_weight::CacheWeight;

pub use crate::util::futures::retry;

pub use crate::util::futures::TimeoutError;

pub use crate::util::stats::MovingStats;

Modules

future
Futures

q

s

stream
Asynchronous streams

Macros

Debug
Derive macro generating an impl of the trait Debug.

anyhow
Construct an ad-hoc error from a string or existing non-anyhow error value.

crit
Log critical level record

debug
Log debug level record

error
Log error level record

info
Log info level record

lazy_static

o
Macro for building group of key-value pairs: OwnedKV

trace
Log trace level record

warn
Log warning level record

Structs

Arc
A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BlockPtr
A block hash and block number from a specific Ethereum block.

BlockState

DataSourceTemplateInfo

Duration
A Duration type to represent a span of time, typically used for system timeouts.

Error
The Error type, a wrapper around a dynamic error type.

EthereumBlock

EthereumBlockWithCalls

EthereumCall

Histogram
A [Metric] counts individual observations from an event or sample stream in configurable buckets. Similar to a Summary, it also provides a sum of observations and an observation count.

HistogramOpts
A struct that bundles the options for creating a Histogram metric. It is mandatory to set Name and Help to a non-empty string. All other fields are optional and can safely be left at their zero value.

HostMetrics

Logger
Logging handle used to execute logging statements

Opts
A struct that bundles the options for creating most [Metric] types.

Pin
A pinned pointer.

Query
A GraphQL query as submitted by a client, either directly or through a subscription.

QueryResult
The result of running a query, if successful.

QueryVariables
Variable values for a GraphQL query.

Registry
A struct for registering Prometheus collectors, collecting their metrics, and gathering them into MetricFamilies for exposition.

SerializableValue
Serializable wrapper around a GraphQL value.

Subscription
A GraphQL subscription made by a client.

Enums

Async
Return type of future, indicating whether a value is ready or not.

AsyncSink
The result of an asynchronous attempt to send a value to a sink.

BigIntSign
A Sign is a BigInt’s composing element.

PrometheusError
The error types for prometheus.

QueryError
Error caused while processing a Query request.

QueryExecutionError
Error caused while executing a Query.

SubgraphVersionSwitchingMode

SubscriptionError
Error caused while processing a Subscription request.

Traits

Collector
An interface for collecting metrics.

Debug
? formatting.

FromIterator
Conversion from an Iterator.

Future
Trait for types which are a placeholder of a value that may become available at some later point in time.

Future01CompatExt
Extension trait for futures 0.1 Future

IntoFuture
Class of types which can be converted into a future.

LightEthereumBlockExt

RuntimeHost
Common trait for runtime host implementations.

RuntimeHostBuilder

Sink
A Sink is a value into which other values can be sent, asynchronously.

Sink01CompatExt
Extension trait for futures 0.1 Sink

Stream
A stream of values, not all of which may have been produced yet.

Stream01CompatExt
Extension trait for futures 0.1 Stream

SubgraphAssignmentProvider
Common trait for subgraph providers.

SubgraphInstanceManager
A SubgraphInstanceManager loads and manages subgraph instances.

SubgraphRegistrar
Common trait for subgraph registrars.

TryFrom
Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryFromValue

TryFutureExt
Adapters specific to Result-returning futures

TryStreamExt
Adapters specific to Result-returning streams

ValueMap

Type Definitions

Counter
A [Metric] represents a single numerical value that only ever goes up.

CounterVec
A Collector that bundles a set of Counters that all share the same [Desc], but have different values for their variable labels. This is used if you want to count the same thing partitioned by various dimensions (e.g. number of HTTP requests, partitioned by response code and method).

DynTryFuture

Gauge
A [Metric] represents a single numerical value that can arbitrarily go up and down.

GaugeVec
A Collector that bundles a set of Gauges that all share the same [Desc], but have different values for their variable labels. This is used if you want to count the same thing partitioned by various dimensions (e.g. number of operations queued, partitioned by user and operation type).

HistogramVec
A Collector that bundles a set of Histograms that all share the same [Desc], but have different values for their variable labels. This is used if you want to count the same thing partitioned by various dimensions (e.g. HTTP request latencies, partitioned by status code and method).

LightEthereumBlock

Poll
Return type of the Future::poll method, indicates whether a future’s value is ready or not.

QueryResultStream
A stream of query results for a subscription.

StartSend
Return type of the Sink::start_send method, indicating the outcome of a send attempt. See AsyncSink for more details.

SubscriptionResult
The result of running a subscription, if successful.

Attribute Macros

async_trait

Derive Macros

Deserialize

Serialize