Function graph::prelude::future::select_all

pub fn select_all<I>(
    iter: I
) -> SelectAll<<<I as IntoIterator>::Item as IntoFuture>::Future> where
    I: IntoIterator,
    <I as IntoIterator>::Item: IntoFuture, 

Creates a new future which will select over a list of futures.

The returned future will wait for any future within iter to be ready. Upon completion or failure the item resolved will be returned, along with the index of the future that was ready and the list of all the remaining futures.

Panics

This function will panic if the iterator specified contains no items.