Module graph::prelude::q

Structs

Directive

Document
Root of query data

Field

FragmentDefinition

FragmentSpread

InlineFragment

Mutation

Number
This represents integer number

ParseError
Error parsing query

Query

SelectionSet

Subscription

VariableDefinition

Enums

Definition

OperationDefinition

Selection

Type

TypeCondition

Value

Traits

Text
Text abstracts over types that hold a string value. It is used to make the AST generic over the string type.

Functions

parse_query
Parses a piece of query language and returns an AST

Type Definitions

Directive

Document

Field

FragmentDefinition

FragmentSpread

InlineFragment

OperationDefinition

Selection

SelectionSet

Type

TypeCondition

Value

VariableDefinition