Module graph::prelude::s

Structs

Directive

DirectiveDefinition

Document

EnumType

EnumTypeExtension

EnumValue

Field

InputObjectType

InputObjectTypeExtension

InputValue

InterfaceType

InterfaceTypeExtension

InvalidDirectiveLocation

ObjectType

ObjectTypeExtension

ParseError
Error parsing schema

ScalarType

ScalarTypeExtension

SchemaDefinition

UnionType

UnionTypeExtension

Enums

Definition

DirectiveLocation

Type

TypeDefinition

TypeExtension

Value

Traits

Text
Text abstracts over types that hold a string value. It is used to make the AST generic over the string type.

Functions

parse_schema
Parses a piece of schema language and returns an AST

Type Definitions

Directive

DirectiveDefinition

Document

EnumType

EnumValue

Field

InputObjectType

InputValue

InterfaceType

ObjectType

ScalarType

Type

TypeDefinition

UnionType

Value