Function graph::prelude::stream::futures_ordered

pub fn futures_ordered<I>(
    futures: I
) -> FuturesOrdered<<<I as IntoIterator>::Item as IntoFuture>::Future> where
    I: IntoIterator,
    <I as IntoIterator>::Item: IntoFuture, 

Converts a list of futures into a Stream of results from the futures.

This function will take an list of futures (e.g. a vector, an iterator, etc), and return a stream. The stream will yield items as they become available on the futures internally, in the order that their originating futures were submitted to the queue. If the futures complete out of order, items will be stored internally within FuturesOrdered until all preceding items have been yielded.

Note that the returned queue can also be used to dynamically push more futures into the queue as they become available.