Struct graph::prelude::stream::Merge

#[must_use = "streams do nothing unless polled"]pub struct Merge<S1, S2> where
    S2: Stream,  {
    stream1: Fuse<S1>,
    stream2: Fuse<S2>,
    queued_error: Option<<S2 as Stream>::Error>,
}

👎 Deprecated:
functionality provided by select now

An adapter for merging the output of two streams.

The merged stream produces items from one or both of the underlying streams as they become available. Errors, however, are not merged: you get at most one error at a time.

Fields
stream1: Fuse<S1>

👎 Deprecated:
functionality provided by select now

stream2: Fuse<S2>

👎 Deprecated:
functionality provided by select now

queued_error: Option<<S2 as Stream>::Error>

👎 Deprecated:
functionality provided by select now