Module graph::runtime

Facilities for creating and reading objects on the memory of an AssemblyScript (Asc) WASM module. Objects are passed through the asc_new and asc_get methods of an AscHeap implementation. These methods take types that implement To/FromAscObj and are therefore convertible to/from an AscType.

Modules

asc_heap

asc_ptr

Structs

AscPtr
A pointer to an object in the Asc heap.

DeterministicHostError

Enums

HostExportError

IndexForAscTypeId

Constants

HEADER_SIZE

Traits

AscHeap
A type that can read and write to the Asc heap. Call asc_new and asc_get for reading and writing Rust structs from and to Asc.

AscIndexId
Marker trait for AssemblyScript types that the id should be in the header.

AscType
A type that has a direct correspondence to an Asc type.

AscValue
An Asc primitive or an AscPtr into the Asc heap. A type marked as AscValue must have the same byte representation in Rust and Asc, including same size, and size must be equal to alignment.

FromAscObj
Type that can be converted from an Asc object of class C.

ToAscObj
Type that can be converted to an Asc object of class C.

TryFromAscObj

Functions

asc_get
Read the rust representation of an Asc object of class C.

asc_new
Instantiate rust_obj as an Asc object of class C. Returns a pointer to the Asc heap.

padding_to_16

try_asc_get