Trait graph::runtime::AscIndexId

pub trait AscIndexId {
    const INDEX_ASC_TYPE_ID: IndexForAscTypeId;
}

Marker trait for AssemblyScript types that the id should be in the header.

Associated Constants

const INDEX_ASC_TYPE_ID: IndexForAscTypeId

Constant string with the name of the type in AssemblyScript. This is used to get the identifier for the type in memory layout. Info about memory layout: https://www.assemblyscript.org/memory.html#common-header-layout. Info about identifier (idof<T>): https://www.assemblyscript.org/garbage-collection.html#runtime-interface

Loading content...