Trait graph::runtime::AscType

pub trait AscType: Sized {
    fn to_asc_bytes(&self) -> Result<Vec<u8>, DeterministicHostError>;

    fn from_asc_bytes(
        asc_obj: &[u8], 
        api_version: Version
    ) -> Result<Self, DeterministicHostError>;

    fn content_len(&self, asc_bytes: &[u8]) -> usize { ... }

    fn asc_size<H: AscHeap + ?Sized>(
        _ptr: AscPtr<Self>, 
        _heap: &H
    ) -> Result<u32, DeterministicHostError> { ... }
}

A type that has a direct correspondence to an Asc type.

This can be derived for structs that are #[repr(C)], contain no padding and whose fields are all AscValue. Enums can derive if they are #[repr(u32)].

Special classes like ArrayBuffer use custom impls.

See https://github.com/graphprotocol/graph-node/issues/607 for more considerations.

Required methods

fn to_asc_bytes(&self) -> Result<Vec<u8>, DeterministicHostError>

Transform the Rust representation of this instance into an sequence of bytes that is precisely the memory layout of a corresponding Asc instance.

fn from_asc_bytes(
asc_obj: &[u8],
api_version: Version
) -> Result<Self, DeterministicHostError>

The Rust representation of an Asc object as layed out in Asc memory.

Loading content...
Provided methods

fn content_len(&self, asc_bytes: &[u8]) -> usize

fn asc_size<H: AscHeap + ?Sized>(
_ptr: AscPtr<Self>,
_heap: &H
) -> Result<u32, DeterministicHostError>

Size of the corresponding Asc instance in bytes. Only used for version <= 0.0.3.

Loading content...