Trait graph::runtime::AscValue

pub trait AscValue: AscType + Copy + Default { }

An Asc primitive or an AscPtr into the Asc heap. A type marked as AscValue must have the same byte representation in Rust and Asc, including same size, and size must be equal to alignment.