Struct graph::Poisoned

pub struct Poisoned<'a, V: 'a, E: 'a + Clone, G: GraphTrait<V=V, E=E> + 'a> { /* fields omitted */ }

Poisoned graph.

The poisoned graph is a graph where some of the edges have been replace with other edges.

Examples

use graph::Graph;
use graph::GraphTrait;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

// Poisoning is meant to work on edges, not nodes. Therefore, all nodes in the
// overriding graph are actually just references to the original graph.
let c = &"C";

use graph::Poisoned;
let poisoned = Poisoned::new(
        &vec![(0, c)],
        &vec![],
        &graph
    ).expect("This does not happen");

// Now one vertex has been replaced:
assert_eq!(poisoned.get(0), Some(&"C"));
assert_eq!(poisoned.get(5), Some(&"B"));

// However...
// Any edge originating from the origin has been overwritten.
assert_eq!(poisoned.get_conn_idval(0).unwrap().count(), 0);

Methods

impl<'a, V: 'a, E: 'a + Clone, G: GraphTrait<V=V, E=E> + 'a> Poisoned<'a, V, E, G>

fn new(vertices: &[(usize, &'a V)],
edges: &[(usize, E, usize)],
origin: &'a G)
-> Result<Poisoned<'a, V, E, G>, Error>

Create a new poisoned graph. See Graph for more detail.