Struct graph::ShortestPathResult

pub struct ShortestPathResult<V> {
    pub path: Vec<NodeIdentifier>,
    pub cost: V,
}

Structure used as the return value of calculating the shortest path

Fields

path

The node identifiers for each node in the calculated path

cost

The cost of the path, calculated as the sum of edge data for each edge on the path