Module graph::task_spawn

Wrapper for spawning tasks that abort on panic, which is our default. The functions in this module should be used to execute futures, serving as a facade to the underlying executor implementation which currently is tokio. This serves a few purposes:

Avoid depending directly on tokio APIs, making upgrades or a potential switch easier.

Reflect our chosen default semantics of aborting on task panic, offering *_allow_panic functions to opt out of that.

Reflect that historically we’ve used blocking futures due to making DB calls directly within futures. This point should go away once https://github.com/graphprotocol/graph-node/issues/905 is resolved. Then the blocking flavors should no longer accept futures but closures.

These should not be called from within executors other than tokio, particularly the blocking functions will panic in that case. We should generally avoid mixing executors whenever possible.

Functions

abort_on_panic

block_on
Runs the future on the current thread. Panics if not within a tokio runtime.

spawn
Aborts on panic.

spawn_allow_panic

spawn_blocking
Aborts on panic.

spawn_blocking_allow_panic
Does not abort on panic, panics result in an Err in JoinHandle.

spawn_thread
Spawns a thread with access to the tokio runtime. Panics if the thread cannot be spawned.