Trait graph::Graph

pub trait Graph<I, D> where I: Eq + Clone + Hash<Hasher> {
    fn add_vertex(&mut self, vertex_id: I, data: D);
    fn add_edge(&mut self, start_vertex: I, end_vertex: I);
    fn add_edge_with_weight(&mut self, start_vertex: I, end_vertex: I, weight: i32);
    fn get_vertex_ids(&self) -> Vec<I>;
    fn get_vertex_data(&self, vertex_id: &I) -> Option<&D>;
    fn get_vertex_neighbours(&self, vertex_id: &I) -> Vec<I>;
    fn get_edge_weight(&self, start_vertex: &I, end_vertex: &I) -> Result<i32, String>;
    fn is_adjacent(&self, start_vertex: &I, end_vertex: &I) -> bool;
    fn is_id_in_graph(&self, vertex_id: &I) -> bool;
    fn vertex_degree(&self, vertex_id: &I) -> Result<u32, String>;

    fn dijkstras_shortest_path(&self, start_vertex: &I, target_vertex: &I) -> Result<GraphPath<I>, String> where Self: Sized { ... }
    fn dijkstras_shortest_paths(&self, start_vertex: &I) -> Result<HashMap<I, GraphPath<I>>, String> where Self: Sized { ... }
    fn diameter_path(&self) -> Result<GraphPath<I>, String> where Self: Sized { ... }
    fn k_core_decomposition(&self) -> HashMap<u32, Vec<I>> where Self: Sized { ... }
}

The Graph trait is used to implement common operations on a graph and provide implementations of graph algorithms that use these operations so that concrete types of graphs can be implemented and the algorithms used on them.

Required Methods

fn add_vertex(&mut self, vertex_id: I, data: D)

The method to add a vertex to the graph.

fn add_edge(&mut self, start_vertex: I, end_vertex: I)

The method to add an edge to the graph between two vertices.

fn add_edge_with_weight(&mut self, start_vertex: I, end_vertex: I, weight: i32)

The method to add an edge to the graph between two vertices, specifying a weight.

fn get_vertex_ids(&self) -> Vec<I>

The method to return a vector of IDs of all vertices in the graph.

fn get_vertex_data(&self, vertex_id: &I) -> Option<&D>

The method to retrieve the data associated with a vertex.

fn get_vertex_neighbours(&self, vertex_id: &I) -> Vec<I>

The method to get the list of vertices that are adjacent to a vertex.

fn get_edge_weight(&self, start_vertex: &I, end_vertex: &I) -> Result<i32, String>

The method to get the weight of an edge between two vertices.

fn is_adjacent(&self, start_vertex: &I, end_vertex: &I) -> bool

The method to check if two vertices are adjacent.

fn is_id_in_graph(&self, vertex_id: &I) -> bool

The method to check if a vertex is in the graph.

fn vertex_degree(&self, vertex_id: &I) -> Result<u32, String>

The method to return the degree of a vertex.

Provided Methods

fn dijkstras_shortest_path(&self, start_vertex: &I, target_vertex: &I) -> Result<GraphPath<I>, String> where Self: Sized

Performs Dijkstra's shortest path algorithm on the graph.

Returns the GraphPath between the two vertices and will end prematurely once the path has been found. Returns an error string if an error occured during the algorithm execution.

This algorithm runs in worst case O(V2) time.

fn dijkstras_shortest_paths(&self, start_vertex: &I) -> Result<HashMap<I, GraphPath<I>>, String> where Self: Sized

Performs Dijkstra's shortest path algorithm on the graph.

Returns a HashMap of target vertices to the GraphPath between the vertex and the target vertex. The HashMap will be empty if an error occured.

This algorithm runs in worst case O(V2) time.

fn diameter_path(&self) -> Result<GraphPath<I>, String> where Self: Sized

Finds the diameter of the graph.

Returns a GraphPath of the path that determined the diameter of the graph.

This uses the dijkstras_shortest_paths function to get all shortest paths pairs and find the longest. This algorithm runs in O(V3) time.

fn k_core_decomposition(&self) -> HashMap<u32, Vec<I>> where Self: Sized

Finds the k core of each vertex in the graph.

Returns a HashMap with the core as a key and a vector of all vertex IDs in that core as a value.

This algorithm runs in O(E) time.