Struct graph::transfo_result::GraphTransformation

#[repr(C)]
pub struct GraphTransformation { /* fields omitted */ }

Structure storing the transformation applied to a graph in a compact way.

Methods

impl GraphTransformation

pub fn new(n: u64) -> GraphTransformation

Constructs a new empty GraphTransformation object with n vertices.

Examples :

use graph::transfo_result::GraphTransformation;
let gt = GraphTransformation::new(0);
assert_eq!(gt.order(),0);
assert_eq!(gt.size(),0);
let gt = GraphTransformation::new(5);
assert_eq!(gt.order(),5);
assert_eq!(gt.size(),0);
for i in 0..5 {
    for j in 0..5 {
        if i == j {
            assert!(gt.is_vertex(i));
            assert!(!gt.is_vertex_modified(i));
        } else {
            assert!(!gt.is_edge(i,j));
            assert!(!gt.is_edge_modified(i,j));
        }
    }
}

pub fn set_name(&mut self, s: String)

pub fn name(&self) -> String

pub fn initial_order(&self) -> u64

Returns the number of vertices of the GraphTransformation before applying the transformations

pub fn order(&self) -> u64

Returns the number of vertices of the GraphTransformation after applying the transformations

pub fn initial_size(&self) -> u64

Returns the number of edges of the GraphTransformation before applying the transformations

pub fn size(&self) -> u64

Returns the number of edges of the GraphTransformation after applying the transformations

pub fn is_edge_modified(&self, i: u64, j: u64) -> bool

Returns true if the edge was removed or added from the initial situation.

pub fn is_edge(&self, i: u64, j: u64) -> bool

Returns true if the graph currently has the given edge. To know if the initial graph had the edge, either it was not modified and the current graph has the edge or it was modified and the edge is absent (e.g., it was removed).

pub fn is_vertex_modified(&self, i: u64) -> bool

Returns true if the vertex was removed or added from the initial situation.

pub fn is_vertex(&self, i: u64) -> bool

Returns true if the graph currently has the given vertex. To know if the initial graph had the vertex, either it was not modified and the current graph has the edge or it was modified and the vertex is absent (e.g., it was removed).

pub fn max_vertex(&self) -> u64

Returns the maximal vertex index of the graph. i.e., if a graph had 0 to 5 vertices and we remove vertex 3, it now has 5 vertices but its maximal index is still 5 since vertex 5 was not removed.

pub fn add_edge(&mut self, i: u64, j: u64)

Adds an edge to the current graph.

Examples :

use graph::transfo_result::GraphTransformation;
use graph::{Graph,GraphNauty};

let mut g = GraphNauty::new(3);
g.add_edge(1,0);
let mut gt: GraphTransformation = (&g).into();
gt.add_edge(1,0);
assert!(gt.is_edge(1,0));
assert!(!gt.is_edge_modified(1,0));
assert_eq!(gt.initial_size(),1);
assert_eq!(gt.size(),1);
gt.add_edge(1, 2);
assert_eq!(gt.initial_size(),1);
assert_eq!(gt.size(),2);
assert!(gt.is_edge(1,2));
assert!(gt.is_edge_modified(1,2));
gt.remove_edge(1,0);
gt.add_edge(1,0);
assert_eq!(gt.initial_size(),1);
assert_eq!(gt.size(),2);
assert!(gt.is_edge(1,0));
assert!(!gt.is_edge_modified(1,0));

pub fn remove_edge(&mut self, i: u64, j: u64)

Adds an edge to the current graph.

Examples :

use graph::transfo_result::GraphTransformation;
use graph::{Graph,GraphNauty};

let mut g = GraphNauty::new(3);
g.add_edge(1,2);
let mut gt: GraphTransformation = (&g).into();
gt.remove_edge(1,0);
assert!(!gt.is_edge(1,0));
assert!(!gt.is_edge_modified(1,0));
assert_eq!(gt.size(),1);
assert_eq!(gt.initial_size(),1);
gt.remove_edge(1, 2);
assert_eq!(gt.size(),0);
assert_eq!(gt.initial_size(),1);
assert!(!gt.is_edge(1,2));
assert!(gt.is_edge_modified(1,2));
gt.add_edge(1,0);
gt.remove_edge(1,0);
assert_eq!(gt.size(),0);
assert_eq!(gt.initial_size(),1);
assert!(!gt.is_edge(1,0));
assert!(!gt.is_edge_modified(1,0));

pub fn add_vertex(&mut self, i: u64)

Adds a vertex to the current graph. The GraphTransformation keeps the same labels for vertices as in the initial graph (except for added vertices whose labels starts at n, the number of vertices in the initial graph). Thus, this new order can be applied to vertices of the initial graph by simply ignoring the vertices added by the transformation.

Examples :

use graph::transfo_result::GraphTransformation;
use graph::{Graph,GraphNauty};

let mut g = GraphNauty::new(0);
let mut gt: GraphTransformation = (&g).into();
gt.add_vertex(0);
assert_eq!(gt.order(), 1);
g = GraphNauty::new(3);
gt = (&g).into();
gt.add_vertex(2);
assert!(gt.is_vertex(2));
assert!(!gt.is_vertex_modified(2));
gt.add_vertex(3);
assert_eq!(gt.order(),4);
assert_eq!(gt.initial_order(),3);
assert!(gt.is_vertex(3));
assert!(gt.is_vertex_modified(3));
gt.remove_vertex(2);
gt.add_vertex(2);
assert!(gt.is_vertex(2));
assert!(!gt.is_vertex_modified(2));

pub fn remove_vertex(&mut self, i: u64)

Removes a vertex to the current graph.

Examples :

use graph::transfo_result::GraphTransformation;
use graph::{Graph,GraphNauty};

let mut g = GraphNauty::new(4);
g.add_edge(2,1);
g.add_edge(2,3);
let mut gt: GraphTransformation = (&g).into();
gt.remove_vertex(2);
assert!(!gt.is_vertex(2));
assert!(gt.is_vertex_modified(2));
for &i in [1,3].iter() {
    assert!(!gt.is_edge(2,i));
    assert!(gt.is_edge_modified(2,i));
}
assert_eq!(gt.order(),3);
assert_eq!(gt.initial_order(),4);
gt.remove_vertex(2);
assert_eq!(gt.order(),3);
assert_eq!(gt.initial_order(),4);
assert!(!gt.is_vertex(2));
assert!(gt.is_vertex_modified(2));
gt.add_vertex(4);
gt.remove_vertex(4);
assert!(!gt.is_vertex(4));
assert!(!gt.is_vertex_modified(4));
assert_eq!(gt.order(),3);
assert_eq!(gt.initial_order(),4);

pub fn initial_graph(&self) -> GraphNauty

Returns the initial graph used to construct the GraphTransformation.

Examples :

use graph::{Graph,GraphNauty,GraphIter};
use graph::transfo_result::GraphTransformation;

let g = GraphNauty::new(0);
let gt: GraphTransformation = (&g).into();
let g = gt.initial_graph();
assert_eq!(g.order(),0);
assert_eq!(g.size(),0);
let mut m = 0;
let mut vi = g.vertices();
while let Some(v) = vi.next() {
    let mut wi = vi.clone();
    for w in wi.filter(|x| g.is_edge(v,*x)) {
        m += 1;
    }
}
assert_eq!(m,0);
let mut g = GraphNauty::new(5);
let edges = [(0,1), (0,2), (3,4)];
for (i,j) in edges.iter() {
    g.add_edge(*i,*j);
}
let mut gt: GraphTransformation = (&g).into();
gt.remove_edge(0,1);
gt.add_edge(1,4);
gt.remove_vertex(1);
gt.add_vertex(5);
let res_g = gt.initial_graph();
assert_eq!(res_g.order(),5);
assert_eq!(res_g.size(),3);
let mut vi = res_g.vertices();
let mut i = 0;
while let Some(v) = vi.next() {
    let mut wi = vi.clone();
    for w in wi.filter(|x| res_g.is_edge(v,*x)) {
        let (ri,rj) = edges[i];
        assert_eq!(v,ri);
        assert_eq!(w,rj);
        i += 1;
    }
}

pub fn final_graph(&mut self) -> GraphNauty

Returns the graph obtained by applying the transformations to the initial graph

Examples :

use graph::{Graph,GraphNauty,GraphIter};
use graph::transfo_result::GraphTransformation;

let g = GraphNauty::new(0);
let mut gt: GraphTransformation = (&g).into();
let g = gt.final_graph();
assert_eq!(g.order(),0);
assert_eq!(g.size(),0);
let m = g.vertices().flat_map(|x| g.neighbors(x).zip(std::iter::repeat(x))).count();
assert_eq!(m,0);
let mut g = GraphNauty::new(5);
let mut edges = vec![(0,1), (0,2), (3,4)];
for (i,j) in edges.iter() {
    g.add_edge(*i,*j);
}
let mut gt: GraphTransformation = (&g).into();
gt.remove_edge(0,1);
gt.add_edge(1,4);
gt.remove_vertex(1);
gt.add_vertex(5);
gt.add_vertex(6);
gt.add_edge(2,6);
edges = vec![(0,1), (1,5), (2,3)];
let res_g = gt.final_graph();
assert_eq!(res_g.order(),6,"graph must have right order");
assert_eq!(res_g.size(),3, "graph must have right size");
let mut edges_iter = edges.iter();
let mut verts = res_g.vertices();
while let Some(i) = verts.next() {
    let mut verts2 = verts.clone();
    while let Some(j) = verts2.next() {
        if res_g.is_edge(i,j) {
            let e = edges_iter.next();
            assert!(e.is_some());
            let (ri,rj) = e.unwrap();
            assert_eq!(i,*ri, "edge must have fist vertex {}",i);
            assert_eq!(j,*rj, "edge must have last vertex {}",j);
        }
    }
}

pub fn canon(&mut self)

Computes the canonical form of the result as well as the order of the vertices in this form.

pub fn canon_order(&mut self) -> Vec<u64>

Computes the canonical form of the result as well as the order of the vertices in this form (if not already computed). And returns the canonical ordering of the vertices.

pub fn tocsv(&self) -> String

pub fn reorder(&self, order: &[u64]) -> Result<GraphTransformation, ()>

Reorder the vertices of a GraphTransformation by using the new order given as parameter. Note that the initial graph is no longer the same since its vertices are reordered.