Type Definition graph::StepUnit

pub type StepUnit<Ix> = Step<(), Ix>;

A stateless step.

The structure comes in handy when working with two Step intances bound to diferent data types for storing state e.g. Step<S, Ix> and Step<Y, Ix>

We then construct a copy of the call-chain removing the state from each step, the final result is two StepUnit<Ix> that we can work with.

Implementations

impl<Ix: IndexType> StepUnit<Ix>

pub fn from_step<S>(step: Rc<Step<S, Ix>>) -> Rc<Self>

pub fn make_void<S>(step: Rc<Step<S, Ix>>) -> Step<(), Ix>

pub fn into_step(step: Rc<StepUnit<Ix>>) -> Step<(), Ix>