Trait graph::util::cache_weight::CacheWeight

pub trait CacheWeight {
    fn indirect_weight(&self) -> usize;

    fn weight(&self) -> usize { ... }
}

Estimate of how much memory a value consumes. Useful for measuring the size of caches.

Required methods

fn indirect_weight(&self) -> usize

The weight of values pointed to by this value but logically owned by it, which is not accounted for by size_of.

Loading content...
Provided methods

fn weight(&self) -> usize

Total weight of the value.

Loading content...