Module graph::util::lfu_cache

Structs

CacheEntry
PartialEq and Hash are delegated to the key.

LfuCache
Each entry in the cache has a frequency, which is incremented by 1 on access. Entries also have a weight, upon eviction first stale entries will be removed and then non-stale entries by order of least frequency until the max weight is respected. This cache only removes entries on calls to evict, so the max weight may be exceeded until evict is called. Every STALE_PERIOD evictions entities are checked for staleness.

Constants

STALE_PERIOD

Type Definitions

Priority