Crate graph

Modules

snap

Structs

Graph

A study area imported from OpenStreetMap.

GtfsModel

Intersection

An intersection between one or more roads. This might represent a dead-end.

IntersectionID

Position

A position along a road, along with the closest intersection

ProfileID

Road

Represents an edge going between exactly two Intersections.

RoadID

Route

A route between two positions.

Router

Manages routing queries for one profile. This structure uses contraction hierarchies to calculate routes very quickly. They are slower to construct, but fast to query.

Timer

Enums

Direction

How can a Road be crossed by a particular profile?

GtfsSource

PathStep

A single step along a route