Module graph::invariants

Module containing implementations of different graph invariants

Enums

Distance

Represents distances that can be infinite.

Functions

avdist

Computes the average distance of the graph.

avecc

Computes the average eccentricity of the graph.

connected_components

Returns the connected components of the graph.

deg_max

Computes the maximum degree in the graph.

deg_min

Computes the minimum degree in the graph.

diameter

Returns the length of the longest shortest path in the graph. This can also be defined as the maximum excentricity.

diametral_paths

Returns the list of all elementary paths in g whose length is maximal.

dnm

Diameter of Hnm with same order and size as the graph given in parameter. Hnm is the graph on n vertices and m edges obtained by taking the biggest clique that does not disconnect the graph and adding a path with the remaining vertices linked to a vertex of the clique. If there are remaining edges, they are added between the closest vertex of the path to the clique and vertices of the clique.

eccentricities

Computes the eccentricities of the graph

eci

Computes the eccentric connectivity index of a graph

floyd_warshall

Implementation of the floyd-warshall algorithm. u64::MAX is used to represent infinity.

irregularity

Computes the irregularity of a graph. This invariant is defined as the sum for all edges of the absolute value of the difference between the degrees of its extremities. If the graph has no edges, 0 is returned.

is_connected

Tests whether the graph is connected. i.e., if each vertex can reach every other vertex.

minus_avecc_avdist

Computes the difference between the average eccentricity and the average distance. This does not reuse avecc and avdist to avoid computing the distance matrix two times instead of one.

num_dom

Computes the number of dominant vertices

num_pending

Computes the number of pending vertices

shortest_paths_from

Returns all the shortests paths from start.

shortests_paths

Computes all shortests paths between every pair of edges.