graph

Struct Graph

pub struct Graph {
    pub roads: Vec<Road>,
    pub intersections: Vec<Intersection>,
    pub mercator: Mercator,
    pub profile_names: BTreeMap<String, ProfileID>,
    pub walking_profile_for_transit: Option<ProfileID>,
    pub routers: Vec<Router>,
    pub boundary_polygon: Polygon,
    pub timestamp: Option<i64>,
    pub gtfs: GtfsModel,
}

A study area imported from OpenStreetMap.

Fields
roads: Vec<Road>intersections: Vec<Intersection>mercator: Mercator

Graph stores all geometry in a Mercator projection for the study area. This field helps translation to/from WGS84.

profile_names: BTreeMap<String, ProfileID>walking_profile_for_transit: Option<ProfileID>routers: Vec<Router>

Per profile

boundary_polygon: Polygon

A polygon covering the study area.

timestamp: Option<i64>

Seconds since the epoch, representing how recent the OSM data is

gtfs: GtfsModel
Implementations

impl Graph

pub fn new<R: OsmReader>( input_bytes: &[u8], osm_reader: &mut R, post_process_graph: Box<dyn Fn(&mut Graph) -> Result<()>>, scrape_graph: Box<dyn Fn(&mut R, &Graph) -> Result<()>>, profiles: Vec<(String, Box<dyn Fn(&Tags, &LineString) -> (Direction, Duration)>)>, timer: &mut Timer, ) -> Result<Graph>

Constructs a graph from OpenStreetMap data.

input_bytes: Bytes of an osm.pbf or osm.xml file

osm_reader: A callback for every OSM element read, to extract non-graph data

post_process_graph: A callback to remove edges and intersections after initially importing.

scrape_graph: A callback to capture anything from osm2graph that’s otherwise lost. It can be stored on the osm_reader struct.

profiles: A list of named profiles. Each one assigns an access direction and cost, given OSM tags and a Euclidean center-line. If every profile assigns Direction::None, then the Road is completely excluded from the graph.

impl Graph

pub fn get_costs( &self, starts: Vec<IntersectionID>, profile: ProfileID, public_transit: bool, start_time: NaiveTime, end_time: NaiveTime, ) -> HashMap<RoadID, Duration>

From a list of start intersections, floods out the graph for a profile until end_time is reached. Returns the time needed to reach each road within that range. This query is not precise about positions along a road.

impl Graph

pub fn snap_route( &self, input: &LineString, profile: ProfileID, ) -> Result<Route>

Given an input LineString (in Mercator), try to snap/map-match it to a given profile’s graph

impl Graph

pub fn transit_route_gj( &self, start: Position, end: Position, debug_search: bool, use_heuristic: bool, start_time: NaiveTime, timer: Timer, ) -> Result<String>

impl Graph

pub fn render_debug(&self) -> FeatureCollection

Returns GeoJSON with roads, intersections, and bus stops

pub fn get_inverted_boundary(&self) -> Result<String>

Return a polygon covering the world, minus a hole for the boundary, in WGS84

pub fn find_edge(&self, i1: IntersectionID, i2: IntersectionID) -> Option<&Road>

Find the Road going from i1 to i2 or vice versa

pub fn snap_to_road(&self, pt: Coord, profile: ProfileID) -> Position

Given a point (in Mercator) and profile, snap to a position along some road that profile can cross.

pub fn add_profile( &mut self, name: String, profile: Box<dyn Fn(&Tags, &LineString) -> (Direction, Duration)>, ) -> ProfileID