Struct graph::Graph

pub struct Graph {
    pub roads: Vec<Road>,
    pub intersections: Vec<Intersection>,
    pub mercator: Mercator,
    pub router: EnumMap<Mode, Router>,
    pub boundary_polygon: Polygon,
    pub gtfs: GtfsModel,
}

A study area imported from OpenStreetMap.

Fields
roads: Vec<Road>intersections: Vec<Intersection>mercator: Mercator

Graph stores all geometry in a Mercator projection for the study area. This field helps translation to/from WGS84.

router: EnumMap<Mode, Router>boundary_polygon: Polygon

A polygon covering the study area.

gtfs: GtfsModel
Implementations

impl Graph

pub fn get_costs( &self, starts: Vec<IntersectionID>, mode: Mode, public_transit: bool, start_time: NaiveTime, end_time: NaiveTime, ) -> HashMap<RoadID, Duration>

From a list of start intersections, floods out the graph for a mode until end_time is reached. Returns the time needed to reach each road within that range. This query is not precise about positions along a road.

impl Graph

pub fn new<F: FnOnce(&mut Vec<Road>), R: OsmReader>( input_bytes: &[u8], osm_reader: &mut R, modify_roads: F, timer: &mut Timer, ) -> Result<Graph>

Constructs a graph from OpenStreetMap data.

input_bytes: Bytes of an osm.pbf or osm.xml file

osm_reader: A callback for every OSM element read, to extract non-graph data

modify_roads: Runs before any routing structures are calculated. Use to modify access per mode.

impl Graph

pub fn snap_route(&self, input: &LineString, mode: Mode) -> Result<Route>

Given an input LineString (in Mercator), try to snap/map-match it to a given Mode’s graph

impl Graph

pub fn transit_route_gj( &self, start: Position, end: Position, debug_search: bool, use_heuristic: bool, start_time: NaiveTime, timer: Timer, ) -> Result<String>

impl Graph

pub fn render_debug(&self) -> FeatureCollection

Returns GeoJSON with roads and stops

pub fn get_inverted_boundary(&self) -> Result<String>

Return a polygon covering the world, minus a hole for the boundary, in WGS84

pub fn find_edge(&self, i1: IntersectionID, i2: IntersectionID) -> Option<&Road>

Find the Road going from i1 to i2 or vice versa

pub fn snap_to_road(&self, pt: Coord, mode: Mode) -> Position

Given a point (in Mercator) and mode, snap to a position along some road that mode can cross.