Struct graph::Graph

pub struct Graph<V, E> { /* fields omitted */ }

Graph structure.

Methods

impl<'a, V: 'a, E: 'a> Graph<V, E>

fn new<NI, EI>(vertices: NI, edges: EI) -> Result<Graph<V, E>, Error> where NI: IntoIterator<Item=(usize, V)>,
EI: IntoIterator<Item=(usize, E, usize)>

Create a new graph, with the given vertices and edges.

Every vertex is of the form (id : usize, v : V), and every edge is of the form (from : usize, e : E, to : usize), which connect the vertex with index from to the vertex with index to.

Returns: A new graph instance, or an error if an edge originates from a nonexisting vertex.

Examples

use graph::Graph;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

fn contains(&self, index: usize) -> bool

Returns whether the graph contains the index.

Examples

use graph::Graph;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

assert_eq!(graph.contains(0), true); // The vertex with index 0 has data "A".
assert_eq!(graph.contains(2), false);

fn get_edges(&'a self, index: usize) -> Option<IterEdges<'a, E>>

Returns all edges connected to a node.

Examples

use graph::Graph;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

let mut edges = graph.get_edges(0).expect("This does not happen");
assert_eq!(edges.next(), Some(&"Edge from A to B"));
assert_eq!(edges.next(), None);

fn get_connids(&'a self, index: usize) -> Option<IterConnIds<'a, E>>

Returns all the nodes this node is connected to.

Examples

use graph::Graph;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

let mut edges = graph.get_connids(0).expect("This does not happen");
assert_eq!(edges.next(), Some(5));
assert_eq!(edges.next(), None);

fn list_ids(&'a self) -> ListIds<'a, V, E>

Returns a list of all possible ids.

Examples

use graph::Graph;
let graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

let mut ids = graph.list_ids();
assert_eq!(ids.next(), Some(0));
assert_eq!(ids.next(), Some(5));
assert_eq!(ids.next(), None);

fn get_edge_mut(&'a mut self, from: usize, to: usize) -> Option<&'a mut E>

Returns a mutable version of the edge between from and to.

Examples

use graph::Graph;
let mut graph = Graph::new(
            vec![(0, "A"), (5, "B")],
            vec![(0, "Edge from A to B", 5)]
    ).expect("This does not happen");

match graph.get_edge_mut(0, 5) {
    Some(t) => *t = "Hello",
    None => panic!("This does not happen!"),
}

let mut edges = graph.get_edges(0).expect("This does not happen");
assert_eq!(edges.next(), Some(&"Hello"));
assert_eq!(edges.next(), None);