Trait graph::Graph

pub trait Graph<N, E> where N: Eq + Clone + Hash<Hasher>, E: Eq + Clone + Hash<Hasher> + Edge<N> {
    fn new() -> Self;
    fn add_node(&mut self, node_id: N);
    fn add_edge(&mut self, source: N, destination: N, weight: i32);
    fn get_nodes(&self) -> Vec<N>;
    fn get_node_neighbours(&self, node_id: &N) -> Vec<N>;
    fn get_edge(&self, source: &N, destination: &N) -> Result<E, String>;
    fn get_edges(&self) -> Vec<E>;
    fn is_adjacent(&self, source: &N, destination: &N) -> bool;
    fn is_node_in_graph(&self, node: &N) -> bool;
    fn degree(&self, node: &N) -> Result<u32, String>;

    fn dijkstras_shortest_path(&self, source: &N, destination: &N) -> Result<GraphPath<N>, String> where Self: Sized { ... }
    fn dijkstras_shortest_paths(&self, source: &N) -> Result<HashMap<N, GraphPath<N>>, String> where Self: Sized { ... }
    fn diameter_path(&self) -> Result<GraphPath<N>, String> where Self: Sized { ... }
    fn k_core_decomposition(&self) -> HashMap<u32, Vec<N>> where Self: Sized { ... }
    fn kruskal_min_spanning_tree(&self) -> Self where Self: Sized { ... }
}

The Graph trait is used to implement common operations on a graph and provide implementations of graph algorithms that use these operations so that concrete types of graphs can be implemented and the algorithms used on them.

Required Methods

fn new() -> Self

Creates a new instance of the graph.

fn add_node(&mut self, node_id: N)

The method to add a node to the graph.

fn add_edge(&mut self, source: N, destination: N, weight: i32)

The method to add an edge to the graph between two nodes, specifying a weight.

fn get_nodes(&self) -> Vec<N>

The method to return a vector of IDs of all nodes in the graph.

fn get_node_neighbours(&self, node_id: &N) -> Vec<N>

The method to get the list of nodes that are adjacent to a node.

fn get_edge(&self, source: &N, destination: &N) -> Result<E, String>

The method to get edge between two nodes.

fn get_edges(&self) -> Vec<E>

The method to get the edge set in the graph.

fn is_adjacent(&self, source: &N, destination: &N) -> bool

The method to check if two nodes are adjacent.

fn is_node_in_graph(&self, node: &N) -> bool

The method to check if a node is in the graph.

fn degree(&self, node: &N) -> Result<u32, String>

The method to return the degree of a node.

Provided Methods

fn dijkstras_shortest_path(&self, source: &N, destination: &N) -> Result<GraphPath<N>, String> where Self: Sized

Performs Dijkstra's shortest path algorithm on the graph.

Returns the GraphPath between the two vertices and will end prematurely once the path has been found. Returns an error string if an error occured during the algorithm execution.

This algorithm runs in worst case O(V2) time.

fn dijkstras_shortest_paths(&self, source: &N) -> Result<HashMap<N, GraphPath<N>>, String> where Self: Sized

Performs Dijkstra's shortest path algorithm on the graph.

Returns a HashMap of target vertices to the GraphPath between the vertex and the target vertex. The HashMap will be empty if an error occured.

This algorithm runs in worst case O(V2) time.

fn diameter_path(&self) -> Result<GraphPath<N>, String> where Self: Sized

Finds the diameter of the graph.

Returns a GraphPath of the path that determined the diameter of the graph.

This uses the dijkstras_shortest_paths function to get all shortest paths pairs and find the longest. This algorithm runs in O(V3) time.

fn k_core_decomposition(&self) -> HashMap<u32, Vec<N>> where Self: Sized

Finds the k core of each vertex in the graph.

Returns a HashMap with the core as a key and a vector of all vertex IDs in that core as a value.

This algorithm runs in O(E) time.

fn kruskal_min_spanning_tree(&self) -> Self where Self: Sized

Creates a minimum spanning tree of the graph using Kruskal's algorithm.

Returns the minimum spanning tree in the same graph implementation that is used.

This algorithm runs in O(Eα(V)) time, where α is the extremely slowly growing inverse of the single-valued Ackermann function.